FLOWFoil

Documentation for FLOWFoil.

FLOWFoil.Freestream

FLOWFoil.Mesh

FLOWFoil.MeshSystem

FLOWFoil.Problem

FLOWFoil.Solution

FLOWFoil.assembleboundaryconditions

FLOWFoil.assemblematrixa

FLOWFoil.assemblematrixa!

FLOWFoil.assemblevortexcoefficients

FLOWFoil.distances

FLOWFoil.generatemesh

FLOWFoil.get_a

FLOWFoil.get_d

FLOWFoil.get_h

FLOWFoil.get_normal

FLOWFoil.get_psibargamma

FLOWFoil.get_psitildegamma

FLOWFoil.get_r

FLOWFoil.get_theta

FLOWFoil.get_vortex_influence

FLOWFoil.joukowsky

FLOWFoil.karmantrefftz

FLOWFoil.normals

FLOWFoil.rawjoukowsky

FLOWFoil.rawkarmantrefftz

FLOWFoil.sizesystem

FLOWFoil.Freestream — Type

Freestream{TF}

Freestream Definition.

Fields:

'reynolds::Vector{Float}' : Reynolds Numbers.

'machs::Vector{Float}' : Mach Numbers.

'anglesofattack::Vector{Float}' : Angles of attack.

FLOWFoil.Mesh — Type

Mesh{TF}

Mesh for single body.

Fields:

'airfoil_nodes::Array{Array{Float,2}}' : [x y] node (panel edge) locations for airfoil

'wake_nodes::Array{Array{Float,2}}' : [x y] node (panel edge) locations for wake

'wake_midpoints::Array{Array{Float,2}}' : [x y] wake panel midpoint locations

'blunt_te::Bool' : boolean for whether or not the trailing edge is blunt or not.

Assuptions:

x and y coordinates start at the bottom trailing edge and proceed clockwise.

x and y coordinates are normalized such that the airfoil has chord length 1.0.

FLOWFoil.MeshSystem — Type

MeshSystem{TF}

System of meshes to solve.

Fields:

'meshes::Array{Mesh}' : Array of mesh objects.

'scales::Vector{Float}' : Airfoil scaling factors.

'angles::Vector{Float}' : Airfoil angles of attack.

'locations::Array{Array{TF}}' : Array of leading edge locations.

FLOWFoil.Problem — Type

Problem{TM,TS,TB}

Problem definition and method selection.

Fields:

'meshsystem::MeshSystem : Mesh System to solve.

'panelshape::String' : Shape of panel to use, e.g., "le.g.r".

'singularitytype::String' : Type of singularity to use, e.g., "source".

'singularityorder::String' : Order of singularity distributions, e.g., "constant".

'verbose::Bool' : Flag to print out verbose statements.

'debug::Bool' : Flag to print out debugging statements.

FLOWFoil.Solution — Type

Solution{TM,TF}

Output object containing solution and useful items.

Fields:

'meshsystem::MeshSystem' : Mesh System used in solution (potentially modified from input meshes).

'strengthsvec::Array{Float,2}' : singularity strengths.

'geocoeffmat::Array{Float,2}' : Geometric Coefficient Matrix used in solution.

'bccoeffvec::Vector{Float}' : Boundary Coefficient Vector used in solution.

'lift::Vector{Float}' : Lift Coefficients.

'drag::Vector{Float}' : Total Drag Coefficients.

'pdrag::Vector{Float}' : Pressure Drag Coefficients.

'idrag::Vector{Float}' : Induced Drag Coefficients.

'moment::Vector{Float}' : Moment Coefficients.

FLOWFoil.assembleboundaryconditions — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.assemblematrixa! — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.assemblematrixa — Method

FLOWFoil.assemblevortexcoefficients — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.distances — Method

distances(meshsystem)

meshsystem::MeshSystem : Mesh System for which to calculate distances.

Returns

rs::Array{Float,2} : Distances from panel edges to control points.

FLOWFoil.generatemesh — Method

generatemesh(coordinates; order="linear")

Create panels from input geometry coordinates.

Arguments:

'x::Vector{Float}' : x coordinates defining airfoil geometry.

'y::Vector{Float}' : y coordinates defining airfoil geometry.

Returns

mesh::Mesh : Geometry mesh, including panel edge points and collocation points.

FLOWFoil.get_a — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.get_d — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.get_h — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.get_normal — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.get_psibargamma — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.get_psitildegamma — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.get_r — Method

function get_r(edge,point)

Calculate the vector, $\mathbf{r}$, and distance, $|r|$, from the node to the evaluation point

Arguments:

node::Array{Float} : [x y] position of node

point::Array{Float} : [x y] position of point.

FLOWFoil.get_theta — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.get_vortex_influence — Method

function_name(args; kwargs)

Function Description.

Detailed Description.

Arguments:

arg::type : description.

Returns

output::type : description.

FLOWFoil.joukowsky — Function

joukowsky(center::Array{Float64}=[-0.1 0.1] ; N::Int=360)

Calculate the unit Joukowsky Airfoil based on a cylinder's center point.

FLOWFoil.karmantrefftz — Function

karmantrefftz(center::Array{Float64}=[-0.1 0.1], wedge::Float64=0.0;N::Int=360)

Calculate geometry of Karman-Trefftz airfoil based on the center point of a cylinder and a wedge angle.

FLOWFoil.normals — Method

FLOWFoil.rawjoukowsky — Function

rawjoukowsky(center::Array{Float64}=[-0.1 0.1] ; N::Int=360)

Calculate geometry of Joukowsky Airfoil based on a cylinder's center point.

FLOWFoil.rawkarmantrefftz — Function

rawkarmanTrefftzFoil(center::Array{Float64}=[-0.1 0.1], wedge::Float64=0.0;N::Int=360)

Calculate geometry of Karman-Trefftz airfoil based on the center point of a cylinder and a wedge angle.

FLOWFoil.sizesystem — Method

sizesystem(meshsystem)

Count size of inviscid system matrix.

Arguments:

'meshsystem::MeshSystem' : The system for which to calculate the linear system size.