Public

FLOWFoil.AirfoilTools.BasicBSpline

FLOWFoil.AirfoilTools.CST

FLOWFoil.AirfoilTools.Joukowsky

FLOWFoil.AirfoilTools.KarmanTrefftz

FLOWFoil.AirfoilTools.KarmanTrefftz

FLOWFoil.AirfoilTools.ModifiedPARSEC

FLOWFoil.AirfoilTools.NACA4

FLOWFoil.AirfoilTools.PARSEC

FLOWFoil.AirfoilTools.basic_bspline

FLOWFoil.AirfoilTools.bernstein

FLOWFoil.AirfoilTools.calculate_modified_parsec_coefficients

FLOWFoil.AirfoilTools.calculate_parsec_coefficients

FLOWFoil.AirfoilTools.cst

FLOWFoil.AirfoilTools.definespline

FLOWFoil.AirfoilTools.determine_cst

FLOWFoil.AirfoilTools.determine_modified_parsec

FLOWFoil.AirfoilTools.determine_naca4

FLOWFoil.AirfoilTools.determine_parsec

FLOWFoil.AirfoilTools.dot

FLOWFoil.AirfoilTools.flip!

FLOWFoil.AirfoilTools.get_spline_coordinates

FLOWFoil.AirfoilTools.half_cst

FLOWFoil.AirfoilTools.joukowsky

FLOWFoil.AirfoilTools.joukowsky_flow

FLOWFoil.AirfoilTools.karman_trefftz

FLOWFoil.AirfoilTools.linear_transform

FLOWFoil.AirfoilTools.modified_parsec

FLOWFoil.AirfoilTools.naca4

FLOWFoil.AirfoilTools.naca4_camber

FLOWFoil.AirfoilTools.naca4_thickness

FLOWFoil.AirfoilTools.norm

FLOWFoil.AirfoilTools.normalize_coordinates!

FLOWFoil.AirfoilTools.parsec

FLOWFoil.AirfoilTools.position_coordinates!

FLOWFoil.AirfoilTools.refine_trailing_edge

FLOWFoil.AirfoilTools.repanel_airfoil

FLOWFoil.AirfoilTools.rotate_coordinates!

FLOWFoil.AirfoilTools.split_cosine_spacing

FLOWFoil.AirfoilTools.split_upper_lower

FLOWFoil.AirfoilTools.whole_cosine_spacing

FLOWFoil.AirfoilTools.z_from_parsec_coefficients

FLOWFoil.AirfoilTools.zero_z_te!

Private

FLOWFoil.AirfoilTools.bernstein — Method

bernstein(r, n, x)

Bernstein Basis Function: binomial(n, r) .* x .^ r .* (1 .- x) .^ (n .- r)

FLOWFoil.AirfoilTools.calculate_modified_parsec_coefficients — Function

calculate_modified_parsec_coefficients(p, uppper_side)

Calculate the PARSEC coefficients using modified parameters (see parsec() docstring) for either the top or bottom curve.

Arguments:

p::NamedTuple : Named tuple of ModifiedPARSEC paramters including:

leading_edge_radius::Float : Leading edge radius

maximum_thickness_x::Float : chordwise position of maximum thickness

maximum_thikcness_z::Float : z-coordinate at maximum thickness

curvature::Float : second derivative of surface geometry at maximum thickness

trailing_edge_tangent::Float : trailing edge tangent angle, radians

trailing_edge_z::Float : z-position of trailing edge

side::Number : +1 for upper side, -1 for lower side

FLOWFoil.AirfoilTools.calculate_parsec_coefficients — Function

calculate_parsec_coefficients(p, side=1)

Calculate the PARSEC coefficients using standard parameters (see parsec() docstring) for either the top or bottom curve.

Arguments:

p::NamedTuple : NamedTuple of PARSEC standard paramters including:

[1]: leading_edge_radius : Leading edge radius

[2]: X : chordwise position of maximum thickness

[3]: Z : z-coordinate at maximum thickness

[4]: curvature : second derivative of surface geometry at maximum thickness

[5]: trailing_edge_angle : trailing edge angle

[6]: boattail_angle : boat-tail angle

[7]: trailing_edge_gap : z-distance between upper and lower surface trailing edge points

[8]: trailing_edge_z : z-position of center of trailing edge

FLOWFoil.AirfoilTools.definespline — Method

definespline()

Calculate the x and y location of the control points, and weight them according to the weighting vector. Also, provide generic knot vectors.

Inputs:

thirdctrlptposition : x position of the 2nd and 6th control points in degree 3.

leadingedgeradius : Leading edge radius

wedge_angle : The trailing edge wedge angle

trailingedgegap : The trailing edge gap

weights : vector weights for the control points, leave empty for unity.

Keyword Arguments:

Outputs:

knots : the knot vector of the curve

controlpoints : control point vector (x,y,w)

FLOWFoil.AirfoilTools.dot — Method

dot(A, B) = sum(a * b for (a, b) in zip(A, B))

A faster dot product.

FLOWFoil.AirfoilTools.get_spline_coordinates — Method

    get_spline_coordinates()

Get plotable coordinates from the weighted control points.

Arguments:

knots::AbstractArray{Float} : knot vector

controlpoints::AbstractArray{Float,2}: Control Point matrix

degree::Int : degree of the NURBS curve

Keyword Arguments:

N::Int : number of coordinates, or panels that will be generated

Returns:

x::AbstractArray{Float} : x Airfoil coordinates

z::AbstractArray{Float} : z Airfoil coordinates

FLOWFoil.AirfoilTools.half_cst — Function

half_cst(coefficients, x=cosine_spacing(N), dz=0.0, N1=0.5, N2=1.0)

Define upper or lower side of airfoil using CST parameterization.

Arguments:

coefficients::AbstractArray{Float} : Vector of CST coefficients.

Keyword Arguments:

dz::Float=0.0 : trailing edge gap

N1::Float=0.5 : Class shape parameter 1

N2::Float=1.0 : Class shape parameter 2

FLOWFoil.AirfoilTools.joukowsky_flow — Method

joukowsky_flow(center, radius, alpha, vinf; N=361)

Calculate the analytic surface velocities and pressures as well as lift coefficient for a joukowsky airfoil.

Arguments:

center::AbstractArray{Float} : [x z] location of circle center relative to origin

radius::Float : Radius of circle

alpha::Float : Angle of attack in degrees

vinf::Float : Freestream velocity

Keyword Arguments:

N::Int : Total number of coordinates to use. Can be even or odd, but it is recommended to be odd for a clear leading edge point.

Returns:

vsurf::AbstractArray{Float} : Magnitude of surface velocities at the nodes

cpsurf::AbstractArray{Float} : Surface pressures at the nodes

cl::Float : Lift coefficient

FLOWFoil.AirfoilTools.linear_transform — Method

linear_transform(range1, range2, values)

Linear transfrom of values from range (source_range[1], source_range[end]) to (target_range[1], target_range[end])

Arguments

source_range::Vector{Float} : range values come from (can also be a Tuple)

target_range::Vector{Float} : range onto which we are transforming (can also be a Tuple)

source_values::Array{Float} : array of source values to transform

Returns

target_values::Array{Float} : array of transformed sourcevalues onto target range

FLOWFoil.AirfoilTools.naca4_camber — Method

naca4_camber(x, max_camber, max_camber_pos)

Compute camber at a given chord-normalized x-position by NACA 4-series camber equations.

Arguments:

x::Float : x position along chordline

max_camber::Float64 : Maximum camber value

max_camber_pos::Float64 : Position of maximum camber

FLOWFoil.AirfoilTools.naca4_thickness — Method

naca4_thickness(x, maxthick; blunt_te=false)

Compute thickness at a given chord-normalized x-position by NACA 4-series thickness equations.

Arguments:

x::Float : x position along chordlin, markersize=3, markershape=:squaree

maxthick::Float : Maximum thickness value

Keyword Arguments:

blunt_te::Bool : Flag whether trailing edge is blunt or not

FLOWFoil.AirfoilTools.norm — Method

norm(A) = sqrt(mapreduce(x -> x^2, +, A))

A faster 2-norm.

FLOWFoil.AirfoilTools.z_from_parsec_coefficients — Function

z_from_parsec_coefficients(a, N::Int=80)

Calculate the x,z airfoil coordinates using the PARSEC polynomial.

Arguments:

a::AbstractArray{Float} : the PARSEC coefficients.