Airfoil Geometry Manipulation Tools

Here we include the variety of methods for manipulating airfoil geometries in useful ways implemented in AirfoilTools.

Deconstruction

FLOWFoil.AirfoilTools.split_upper_lower — Function

split_upper_lower(x, z; idx::Integer=nothing)

Split the upper and lower halves of the airfoil coordinates.

Assumes leading edge point is at first minimum x value if idx is not provided. Returns the upper and lower coordinates each with the leading edge point. Assumes airfoil is defined clockwise starting at the trailing edge.

Arguments

x::AbstractArray{Float} : Vector of x coordinates

z::AbstractArray{Float} : Vector of z coordinates

Keyword Arguments

idx::Integer : optional index at which to split the coordinates

Returns

xl::AbstractArray{Float} : Vector of lower half of x coordinates

xu::AbstractArray{Float} : Vector of upper half of x coordinates

zl::AbstractArray{Float} : Vector of lower half of z coordinates

zu::AbstractArray{Float} : Vector of upper half of z coordinates

split_upper_lower(coordaintes; idx::Integer=nothing)

Split the upper and lower halves of the airfoil coordinates.

Assumes leading edge point is at first minimum x value if idx is not provided. Returns the upper and lower coordinates each with the leading edge point. Assumes airfoil is defined clockwise starting at the trailing edge.

Arguments

coordinates::Matrix{Float} : Matrix of [x z] coordinates

Keyword Arguments

idx::Integer : optional index at which to split the coordinates

Returns

xl::AbstractArray{Float} : View of lower half of x coordinates

xu::AbstractArray{Float} : View of upper half of x coordinates

zl::AbstractArray{Float} : View of lower half of z coordinates

zu::AbstractArray{Float} : View of upper half of z coordinates

Transformation

FLOWFoil.AirfoilTools.flip! — Function

flip!(x)

Flips one dimension of airfoil coordinates.

Moves airfoil left (x) or down (z) by maximum x or z coordinate then flips about the z or x axis, respectively.

Arguments

x::AbstractArray{Float} : vector of x or z coordinates

FLOWFoil.AirfoilTools.zero_z_te! — Function

zero_z_te!(coordinates)

Places trailing edge on the x-axis.

Arguments

coordinates::Array{Float} : Array of [x z] coordinates to be updated in place.

FLOWFoil.AirfoilTools.rotate_coordinates! — Function

rotate_coordinates!(coordinates, angle; rotation_point=[0.0; 0.0])

Rotate coordiantes clockwise about rotation_point by angle in degrees.

Arguments

coordinates::Array{Float} : Array of [x z] coordinates to be updated in place.

angle::Float=0.0 : Angles, in degrees, by which to rotate the coordinates clockwise (positive angle will pitch airfoil up).

Keyword Arguments

rotation_point::AbstractArray{Float}=[0.0; 0.0] : Array of [x z] position of point about which to perform rotation.

FLOWFoil.AirfoilTools.normalize_coordinates! — Function

normalize_coordinates!(coordinates)

Normalize airfoil to unit chord and shift leading edge to zero. Adjusts coordinates in place.

Arguments

coordinates::AbstractArray{Float} : Array of [x z] coordinates

normalize_coordinates!(x, z)

Normalize airfoil to unit chord and shift leading edge to zero. Adjusts coordinates in place.

Arguments

x::Array{Float} : Array of x coordinates

z::Array{Float} : Array of z coordinates

FLOWFoil.AirfoilTools.position_coordinates! — Function

position_coordinates!(coordinates, scale, angle, location)

Scale, Rotate, and Transform (in that order) airfoil coordinates.

Arguments

coordinates::Array{Float} : Array of [x z] coordinates to be updated in place.

Keyword Arguments

scale::Float=1.0 : Value by which to scale coordinates.

angle::Float=0.0 : Angles, in degrees, by which to rotate the coordinates clockwise (positive angle will pitch airfoil up).

location::AbstractArray{Float}=[0.0; 0.0] : Array of [x z] position of leading edge location.

rotation_point::AbstractArray{Float}=[0.0; 0.0] : Array of [x z] position of point about which to perform rotation.

flipped::Bool : flag whether to flip airfoil upside down.

Returns

x::Array{Float} : array of x-coordinates

z::Array{Float} : array of z-coordinates

position_coordinates!(
    coordinates::Vector{AbstractArray{TF}};
    scales=[1.0],
    angles=[0.0],
    locations=[[0.0; 0.0],],
    rotation_points=[[0.0; 0.0],],
    flipped=[false],
) where {TF}

Multi-airfoil version of position_coordinates!.

If keyword arguments are give as single valued vectors, the same values are used for all coordinate sets. If keyword arguments are provided as vectors of length greater than 1, they must have the same length as the set of coordinates. For example, if scaling 3 airfoils, there will be a vector of 3 airfoil coordinate sets input and scales must either be a one element vector or a vector of length 3.

Re-definition

FLOWFoil.AirfoilTools.whole_cosine_spacing — Function

whole_cosine_spacing(N::Integer=160)

Returns cosine spaced x coordinates from 1 to 0 back to 1.

Arguments

N::Integer : Total number of points is N+1.

Returns

x::AbstractArray{Float} : cosine spaced x-coordinates, starting at 1.0, going to 0.0, then back to 1.0.

FLOWFoil.AirfoilTools.split_cosine_spacing — Function

split_cosine_spacing(N::Integer=80)

Returns cosine spaced x coordinates from 0 to 1.

Arguments

N::Integer : Number of points.

Returns

x::AbstractArray{Float} : cosine spaced x-coordinates, starting at 0.0 ending at 1.0.

FLOWFoil.AirfoilTools.repanel_airfoil — Function

repanel_airfoil(x, z; N=160)

Repanels airfoil coordinates using Akima splines with N coordinate points.

Arguments

x::AbstractArray{Float} : vector containing the x coordinates of the airfoil

z::AbstractArray{Float} : vector containing the z coordinates of the airfoil

Keyword Arguements

N::Int : Number of data points to be returned after repaneling. Will only return odd numbers, if N is even, N+1 points will be returned.

Returns

repaneled_x::AbstractArray{Float} : Repaneled, cosine spaced x corrdinates of the airfoil

repaneled_z::AbstractArray{Float} : z coordinates of the repaneled airfoil obtained using an akima spline

repanel_airfoil(coordinates; N=160)

Repanels airfoil coordinates using Akima splines with N coordinate points.

Arguments

coordinates::Array{Float} : Array of [x z] coordinates

Keyword Arguements

N::Int=160 : Number of data points to be returned after repaneling. Will only return odd numbers, if N is even, N+1 points will be returned.

Returns

repaneled_coordinates::Array{Float} : new coordinate array.

FLOWFoil.AirfoilTools.refine_trailing_edge — Function

refine_trailing_edge(coordinates)

Adds points along the trailing edge of an airfoil.

Arguments

coordinates::Array{Float} : Array of [x z] coordinates

Returns

x::AbstractArray{Float} : vector containing the x coordinates of an airfoil with points along the trailing edge.

z::AbstractArray{Float} : vector containing the z coordinates of an airfoil with points along the trailing edge.

refine_trailing_edge(x, z, N::Integer)

Adds points along the trailing edge of an airfoil

Arguments

x::AbstractArray{Float} : vector containing the x coordinates of an airfoil.

z::AbstractArray{Float} : vecotr containing the z coordinate of an airfoil.

N::Integer : Number of data points to be added at the trailing edge.

Returns

x::AbstractArray{Float} : vector containing the x coordinates of an airfoil with points along the trailing edge.

z::AbstractArray{Float} : vector containing the z coordinates of an airfoil with points along the trailing edge.

Contributing

We welcome the addition of more convenience functions for airfoil geometry manipulation.