Basic Tutorials

FLOWFoil includes various panel method implementations that are available based on the method keyword argument. Here we go over the available methods and their various options

Xfoil Method

An Xfoil-like method, actually based on mfoil can be accessed using the Mfoil method type:

FLOWFoil.Mfoil — Type

Mfoil <: Method

Fields:

viscous::Bool : Flag whether to run an inviscid (false) or viscous (true) analysis.

Note that we have also set Xfoil=Mfoil so you can also use the Xfoil method type with identical results. Currently, this method only includes the inviscid parts of Xfoil/Mfoil.

using FLOWFoil

#setup x and y coordinates for the airfoil
x, y = AirfoilTools.naca4()

#create a range of angles of attack
angles_of_attack = range(-5.0, 15.0, step=1)

#solve for invsicid case
viscous = false

#set up Mfoil method
method = Mfoil(viscous)

#solve for outputs
outputs = analyze(x, y, angles_of_attack; method=method)

FLOWFoil.outputs{Matrix{Float64}, Vector{Float64}}([-0.692908072768126 -0.6934105709397785 … -0.6665000542362963 -0.6630363805906164; -0.7421458177430701 -0.7424731035417773 … -0.7099266288461078 -0.7060179988578108; … ; 0.7437749192650895 0.7445314515350743 … 0.7194790365407877 0.7159657903190176; 0.692908072768126 0.6934105709397785 … 0.6665000542362963 0.6630363805906164], [0.5198784026927613 0.5191817801089704 … 0.5557776777030141 0.5603827580132952; 0.44921958520646976 0.44873369051704126 … 0.49600418165520066 0.5015385852888123; … ; 0.4467988694722095 0.44567291767507533 … 0.48234991597833987 0.4873929870928645; 0.5198784026927613 0.5191817801089704 … 0.5557776777030141 0.5603827580132952], [-0.34847343924627516, -0.22779847891023397, -0.10705614988471084, 0.013717082194117653, 0.1344847469562604, 0.2552103795346293, 0.3758575307896451, 0.4963897775210392, 0.6167707326661785, 0.7369640554842715  …  0.9766427337866203, 1.0960557308459074, 1.215136398987623, 1.3338487813045354, 1.4521570279843254, 1.5700254063770003, 1.6874183110429035, 1.804300273780559, 1.9206359736335232, 2.0363902468753956], [0.0005166916701355751, 0.00048227845938455914, 0.00045365837445987957, 0.00043093776021815345, 0.0004142046156752168, 0.00040352834497148, 0.00039895955996946603, 0.0004005299350271381, 0.00040825211434695894, 0.0004221196721562829  …  0.000468170001920482, 0.0005002449549158945, 0.0005382499384007293, 0.0005820844281519555, 0.000631629696558554, 0.0006867491376132376, 0.0007472886415855602, 0.0008130770183491654, 0.0008839264682029976, 0.0009596330988951767], [-0.048888114413008436, -0.05019850205092104, -0.05153111561359062, -0.05288433151666448, -0.054256501075011634, -0.05564595251139171, -0.05705099299325813, -0.05846991069521359, -0.05990097688460577, -0.061342448027721826  …  -0.06424956979577882, -0.06571167854065323, -0.06717711279435354, -0.06864408715097446, -0.07011081432823246, -0.07157550734499417, -0.07303638169843726, -0.07449165753819251, -0.07593956183481718, -0.07737833053995849])

Lewis' Method for Axisymmetric Bodies

A axisymmetric method based on that described by Lewis can be accessed using the Lewis method type:

FLOWFoil.Lewis — Type

Lewis <: Method

Fields:

body_of_revolution::AbstractVector{Bool} : Flag(s) whether bodies are bodies of revolutions or not (false indicates an annular airfoil)

Note that if multiple bodies are used, the annular airfoil should come before the body of revolution.

using FLOWFoil

#setup airfoil coordinates
x, r = AirfoilTools.naca4()

# give the duct some diameter
r .+= 1.0

# indicate that the body is not a body of revolution (lying on the axis)
method = Lewis(body_of_revolution = [false])

# angle of attack defaults to zero, which is what we want for the axisymmetric case
outputs = analyze(x, r; method=method)

FLOWFoil.outputs{Matrix{Float64}, Float64}([-0.7169024195697359; -0.7767614729891681; … ; 0.7525392894347234; 0.7169024195697359;;], [0.4860509208150584; 0.3966416140796979; … ; 0.4336846178570817; 0.4860509208150584;;], 0.0, 0.0, 0.0)

Note

No part of the geometry for an axisymmetric body can reside below z=0, otherwise an error will be thrown.

Martensen's Method for Periodic Bodies (Cascades)

A periodic method for cascade analysis based on that developed by Martensen can be accessed using the Martensen method type:

FLOWFoil.Martensen — Type

Martensen{} <: Method

Fields:

cascade::Bool=true : flag to apply cascade treatment or not

solidity::Float=0.0 : Ratio between airfoil chord length and pitch. Airfoil pitch is simply the distance between chordlines in the cascade.

stagger::Float=0.0 : Angle (in degrees) from axis of ration to airfoil chordline. Note that stagger is equivalent to the inflow angle minus the angle of attack.

transition_value::Float=Inf : pitchtochord ratio at which we stop applying cascade effects (Lewis uses 30 in his implementation)

curvature_correction::Bool=false : flag to apply curvature correction from Lewis

using FLOWFoil

#setup airfoil coordinates
x, y = AirfoilTools.naca4(6,4,12)

#specify range of angle of attack
inflow_angles = range(-5.0, 15.0, step=1)

#solve for airfoil cascade
cascade = true

#setup solidity - only important in cascade case. If cascade = false, then value of solidity will not effect the outputs
solidity = 1.0

#setup airfoil stagger (inflow angle minus angle of attack). When in doubt set this to 0.0
stagger = 0.0

#Value of solidity that the cascade method will transition to solving using the planar method (ie single airfoil instead of a cascade). In this case it is used for blending the cascade solution and planar solution together when the solidity is close to the transition value. When in doubt set this to 1e-4.
transition_value = 0.0001

#curvature correction is currently not working, always set to false
curvature_correction = false

#specify Martensen method
method = Martensen(cascade,solidity,stagger, transition_value, curvature_correction)

#solve for outputs
outputs = analyze(x, y, inflow_angles; method=method)

FLOWFoil.outputs{Matrix{Float64}, Vector{Float64}}([-0.752715994610868 -0.7536519165558849 … -0.7313048403647816 -0.7279098941408012; -0.8423714566255355 -0.8433951699228668 … -0.8179681883698112 -0.8141465446208617; … ; 0.760737016561168 0.7617063280459644 … 0.7395345243159702 0.7361254751574742; 0.8423714566255355 0.8433951699228668 … 0.8179681883698112 0.8141465446208617], [0.4334186314569717 0.4320087886716415 … 0.46519323045904126 0.4701471860119276; 0.2904103290625736 0.2886845873507786 … 0.330928042815009 0.3371654038819113; … ; 0.42127919163361327 0.4198034698147337 … 0.4530886873447516 0.45811928482418285; 0.2904103290625736 0.2886845873507786 … 0.330928042815009 0.3371654038819113], [0.08676891413672283, 0.12084028713833762, 0.15487983933782545, 0.18890021841710058, 0.22291404394050832, 0.2569339181239175, 0.290972436519059, 0.3250421985896213, 0.35915581815446423, 0.393325933671589  …  0.4618863899472196, 0.4963022205491781, 0.5308255457398333, 0.5654692736691848, 0.6002463936374125, 0.6351699842486588, 0.6702532210538034, 0.7055093836085804, 0.740951861863403, 0.776594161789949], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

Hess-Smith 2D Method for Educational Purposes

We also have a version of the Hess-Smith method primarily for educational use that can be accessed with the HessSmith method type:

FLOWFoil.HessSmith — Type

HessSmith <: Method

Fields:

V_inf::TF=1.0 : magnitude of the free stream velocity

using FLOWFoil

#specify airfoil coordinates
x, y = AirfoilTools.naca4(6,4,12)

#specify range of angle of attack
angles_of_attack = range(-5.0, 15.0, step=1)

#specify magnitude of the free stream velocity
vinf = 1.0

#specify HessSmith method
method = HessSmith(vinf)

#solve for outputs
outputs = analyze(x, y, angles_of_attack; method=method)

FLOWFoil.outputs{Matrix{Float64}, Vector{Float64}}([-0.6928965848849973 -0.6678729918939355 … -0.19030941783279443 -0.16278294058994602; -0.7587243723935129 -0.7457778957578849 … -0.4780319591315523 -0.4615162478224243; … ; 0.7980951718897481 0.8101154239291226 … 0.9811146838051676 0.9879108231417423; 0.7587316735985412 0.785618042952302 … 1.2211058180608445 1.2421395883721607], [0.5198943226547077 0.5539456666986432 … 0.9637823254841429 0.9735017142528901; 0.42433732673607005 0.44381533019894137 … 0.7714854460488499 0.7870027529959107; … ; 0.36304409660627346 0.343712999912138 … 0.037413977221885886 0.024032205519405214; 0.4243262474783567 0.38280429058779497 … -0.49109941890204434 -0.5429107570013607], [0.14989169906803065, 0.27476990504291193, 0.3995972610813429, 0.5243358396300218, 0.6489476498817178, 0.7733946493318221, 0.8976387555981775, 1.0216418584999654, 1.1453658323906526, 1.2687725487392527  …  1.514481757437191, 1.6367080948753079, 1.7584648917356853, 1.879714201980454, 2.000418156975706, 2.1205389795891896, 2.2400389984646676, 2.358880662460992, 2.477026555243374, 2.594439410013774], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])