Public API

Analyze

FLOWFoil.analyze — Function

analyze(coordinates, flow_angles=0.0, reynolds=1e6, mach=0.0; method::Method=Mfoil())
analyze(x, y, flow_angles=0.0, reynolds=1e6, mach=0.0; method::Method=Mfoil())

Convenience function for setting up, solving, and post-processing airfoils and airfoil systems.

Arguments

coordinates::NTuple{Matrix{Float}} : Tuple of [x y] matrices of airfoil coordinates (may be a single matrix as well)

flow_angles::Vector{Float} : Vector of angles of attack (may be a single float as well)

OR

x::Vector{Float} : Vector of x-coordinates of airfoil geometry

y::Vector{Float} : Vector of y-coordinates of airfoil geometry

flow_angles::Vector{Float} : Vector of angles of attack (may be a single float as well)

Note that inputting separate vectors for airfoil coordinates is only available for analysis of single airfoils/bodies. Multi-airfoil/body systems require the use of a tuple of matrices for coordinate inputs.

Keyword Arguments

method::MethodType : desired method for solving

Returns

outputs::OutputType : outputs object (note that only inviscid methods are currently implemented)

Methods

FLOWFoil provides access to the following methods.

FLOWFoil.Mfoil — Type

Mfoil <: Method

Fields

viscous::Bool : Flag whether to run an inviscid (false) or viscous (true) analysis.

NOTE: viscous method not yet implemented.

FLOWFoil.Lewis — Type

Lewis <: Method

Fields

body_of_revolution::AbstractVector{Bool} : Flag(s) whether bodies are bodies of revolutions or not (false indicates an annular airfoil)

Note that if multiple bodies are used, the annular airfoil should come before the body of revolution.

FLOWFoil.Martensen — Type

Martensen <: Method

Fields

cascade::Bool=true : flag to apply cascade treatment or not

solidity::Float=0.0 : Ratio between airfoil chord length and pitch. Airfoil pitch is simply the distance between chordlines in the cascade.

stagger::Float=0.0 : Angle (in degrees) from axis of ration to airfoil chordline. Note that stagger is equivalent to the inflow angle minus the angle of attack.

transition_value::Float=Inf : pitchtochord ratio at which we stop applying cascade effects (Lewis uses 30 in his implementation)

curvature_correction::Bool=false : flag to apply curvature correction from Lewis

FLOWFoil.HessSmith — Type

HessSmith <: Method

Fields

V_inf::TF=1.0 : magnitude of the free stream velocity

FLOWFoil.LegacyXfoil — Type

LegacyXfoil <: Method

In general usage, this struct is generated by calling the LegacyXfoil constructor.

Fields

reynolds: Reynolds number

mach: Mach number

xtrip: forced transition x/c locations on top/bottom sides

reinit: reinitialize the solution? (rather than use the previous solution)

iter: Maximum iterations for viscous analyses

npan: Number of panels

percussive_maintenance: Call do_percussive_maintenance upon convergence failure?

printdata: Print data obtained from XFOIL during the solution?

zeroinit: Start angle of attack sweeps from zero? If true, results will be sorted by ascending angle of attack.

clmaxstop: Stop if lift coefficient decreases twice consecutively going up?

clminstop: Stop if lift coefficient increases twice consecutively going down?

FLOWFoil.NeuralFoil — Type

NeuralFoil <: Method

In general usage, this struct is generated by calling the NeuralFoil constructor.

Fields:

Re::Float

Ma::Float

n_crit::Float

xtr_upper::Float

xtr_lower::Float

net_cache::NeuralFoil.NetParameters

Outputs

Based on the method used, FLOWFoil returns one of the following types:

FLOWFoil.InviscidOutputs — Type

InviscidOutputs

Note: not all methods return values for all outputs. Methods will return zeros in such cases.

Fields

vs: surface velocities on each body, nominally a matrix, but becomes a vector of matrices in the multi-body case with dimensions [body][panel,angle]

cp: pressure coefficient for each panel of each body, becomes a vector of matrices in the multi-body case with dimensions [body][panel,angle]

cl: lift coefficient of each body, nominally a vector but becomes a matrix in the multi-body case with dimensions angle x body

cd: total drag coefficient of each body, but becomes a matrix in the multi-body case with dimensions angle x body

cm: moment coefficient of each body, but becomes a matrix in the multi-body case with dimensions angle x body

FLOWFoil.LegacyXFOutputs — Type

LegacyXFOutputs

Fields

cl::Vector{Float}

cd::Vector{Float}

cdp::Vector{Float}

cm::Vector{Float}

converged::Vector{Bool}

Or in the case of NeuralFoil, the native NeuralFoil output (NeuralOutputs) type will be returned