Quick Start

Running FLOWFoil can be done simply with a single method: analyze

FLOWFoil.analyze — Function

analyze(coordinates, flow_angles=0.0, reynolds=1e6, machs=0.0; method::Method=Mfoil())
analyze(x, y, flow_angles=0.0, reynolds=1e6, machs=0.0; method::Method=Mfoil())

Convenience function for setting up, solving, and post-processing airfoils and airfoil systems.

Required Arguments:

coordinates::NTuple{Matrix{Float}} : Tuple of [x y] matrices of airfoil coordinates (may be a single matrix as well)

OR

x::Vector{Float} : Vector of x-coordinates of airfoil geometry

y::Vector{Float} : Vector of y-coordinates of airfoil geometry

Note that inputting separate vectors for airfoil coordinates is only available for analysis of single airfoils/bodies. Multi-airfoil/body systems require the use of a tuple of matrices for coordinate inputs.

Optional Arguments:

flow_angles::Vector{Float}=0.0 : Vector of angles of attack (may be a single float as well)

reynolds::Vector{Float}=1e-6 : Vector of reynolds numbers (may be a single float as well)

machs::Vector{Float}=0.0 : Vector of machs numbers (may be a single float as well)

Note that Reynolds and Mach numbers are only used for viscous methods, and Flow Angles are unused in the axisymmetric methods.

Keyword Arguments:

method::Method=Mfoil() : desired method for solving

Returns:

outputs::NTuple : named tuple with outputs. Nominally contains

cl: lift coefficient of each body

cd: total drag coefficient of each body

cdp: profile drag coefficient of each body

cm: moment coefficient of each body

tangential_velocities: surface velocities on each body

surface_pressures: surface pressures on each body

convergenced: convergence flag

auxiliary outputs: a named tuple that contains additional outputs applicable to the method used.

As an introductory example, we will do a quick analysis of a NACA 2412 airfoil, with coordinates from one of the available methods in the Airfoil Tools sub-module.

Note

For any airfoil coordinate generation method (from FLOWFoil or otherwise), the coordinates must start at the trailing edge, and proceed clockwise around the airfoil. For axisymmetric bodies of revolution, coordiantes should also proceed clockwise, starting at the leading edge.

using FLOWFoil

# 2412 is default
x, y = AirfoilTools.naca4()

# choose one or more angles of attack
angles_of_attack = range(-5.0, 15.0, step=1)

outputs = analyze(x, y, angles_of_attack)