FLOWFoil

Documentation for FLOWFoil.

FLOWFoil.BodyMesh

FLOWFoil.Debug

FLOWFoil.InviscidSolution

FLOWFoil.InviscidSystem

FLOWFoil.Parameters

FLOWFoil.Polar

FLOWFoil.Problem

FLOWFoil.ViscousSolution

FLOWFoil.ViscousSystem

FLOWFoil.WakeMesh

FLOWFoil.assemble_boundary_conditions

FLOWFoil.assemble_vortex_coefficients

FLOWFoil.assemble_vortex_matrix

FLOWFoil.camber

FLOWFoil.defaultparameters

FLOWFoil.generate_mesh

FLOWFoil.generate_mesh

FLOWFoil.get_a

FLOWFoil.get_d

FLOWFoil.get_distances

FLOWFoil.get_h

FLOWFoil.get_inviscid_system

FLOWFoil.get_normal

FLOWFoil.get_orientation

FLOWFoil.get_psibargamma

FLOWFoil.get_psibarsigma

FLOWFoil.get_psitildegamma

FLOWFoil.get_psitildesigma

FLOWFoil.get_r

FLOWFoil.get_source_influence

FLOWFoil.get_tangent

FLOWFoil.get_theta

FLOWFoil.get_theta

FLOWFoil.get_vortex_influence

FLOWFoil.initalize_viscous

FLOWFoil.initialize_wake

FLOWFoil.joukowsky

FLOWFoil.karmantrefftz

FLOWFoil.naca4

FLOWFoil.rawjoukowsky

FLOWFoil.rawkarmantrefftz

FLOWFoil.solve

FLOWFoil.solve_inviscid

FLOWFoil.solve_inviscid_system

FLOWFoil.solve_viscous

FLOWFoil.thickness

FLOWFoil.BodyMesh — Type

BodyMesh{TF,TB}

Mesh for single body.

Fields:

'airfoil_nodes::Array{Array{Float,2}}' : [x y] node (panel edge) locations for airfoil

'blunt_te::Bool' : boolean for whether or not the trailing edge is blunt or not.

Assuptions:

x and y coordinates start at the bottom trailing edge and proceed clockwise.

FLOWFoil.Debug — Type

Debug{TIS,TVS}

Fields:

'isystem::InviscidSystem' : Inviscid System Object.

'vsystem::ViscousSystem' : Viscous System Object.

FLOWFoil.InviscidSolution — Type

InviscidSolution{TM,TF,TD}

Fields:

'mesh::BodyMesh' : BodyMesh object describing airfoil nodes etc.

'panelgammas::Array{Float,2}' : $\gamma_0$ and $\gamma_{90}$ values at each airfoil node.

'psi0::Array{Float}' : $\Psi_0$ (constant stream function) 0 and 90 values.

'debug::Debug' : Debug object (or nothing) depending on debug flag in Problem object.

FLOWFoil.InviscidSystem — Type

InviscidSystem{TF}

Fields:

'vcoeffmat::Array{Float,2}' : Vortex Coefficient Matrix used in solution.

'bccoeffvec::Array{Float,2}' : Boundary Coefficient Vector used in solution.

FLOWFoil.Parameters — Type

Parameters{TF}

Solver Parameters.

Fields:

`gamma_air::Float = 1.4' : ratio of specific heats for air

`eta_crit::Float = 9.0' : critical amplification factor

`eta_D::Float = 0.9' : wall/wake dissipation length ratio

`GA::Float = 6.7' : G - Beta locus A constant

`GB::Float = 0.75' : G - Beta locus B constant

`GC::Float = 18.0' : G - Beta locus C constant

`Klag::Float = 5.6' : shear lag constant

`Ctau::Float = 1.8' : shear stress initialization constant

`Etau::Float = 3.3' : shear stree initialization exponent

`rSu::Float = 0.35' : Sutherland temperature ratio

`fw::Float = 2.5' : wake gap continuation factor

`dw::Float = 1.0' : wake length, in airfoil chords

`epsilonw::Float = 1e-5' : first wake point offset, in airfoil chords

FLOWFoil.Polar — Type

Polar{TF}

Fields:

'lift::Float' : Lift Coefficient.

'drag::Float' : Total Drag Coefficient.

'pdrag::Float' : Pressure Drag Coefficient.

'idrag::Float' : Induced Drag Coefficient.

'moment::Float' : Moment Coefficient.

'surfacevelocity::Vector{Float}' : surface velocity distribution

'surfacepressure::Vector{Float}' : surface pressure distribution

FLOWFoil.Problem — Type

Problem{TF,TB}

Problem definition (geometry, operating point(s), and method selection) and output behavior.

Fields:

'coordinates::Array{Float}' : x,y airfoil coordinates.

'angleofattack::Float' : angle of attack to analyze.

'reynolds::Float' : Reynolds number to analyze.

'mach::Float' : Mach number to analyze.

'viscous::Bool' : Flag to solve viscous or inviscid only

'verbose::Bool' : Flag to print out verbose statements

'debug::Bool' : Flag to save the system structs, etc.

FLOWFoil.ViscousSolution — Type

ViscousSolution{}

Fields:

'panelgammas::Array{Float,2}' : $\gamma_0$ and $\gamma_{90}$ values at each airfoil node.

'panelsources::Array{Float,2}' : source values at each airfoil node.

'wakesources::Array{Float,2}' : source values at each wake node.

'psi0::Array{Float}' : $\Psi_0$ (constant stream function) 0 and 90 values.

FLOWFoil.ViscousSystem — Type

Fields:

FLOWFoil.WakeMesh — Type

WakeMesh{TF}

Fields:

'wake_nodes::Array{Float,2}' : x,y wake panel node locations.

'wake_midpoints::Array{Float,2}' : x,y wake panel center point locations.

FLOWFoil.assemble_boundary_conditions — Method

assemble_boundary_conditions(mesh, freestream)

Assemble boundary condition vector.

Arguments:

'mesh::BodyMesh' : mesh system for which to solve

Returns

output::type : description.

FLOWFoil.assemble_vortex_coefficients — Method

assemble_vortex_coefficients(mesh)

Assemble matrix of vortex strength coefficients.

This function only assembles the NxN portion of the influence coefficient matrix. It does not include the kutta condition. Also note that multibody capabilities have not yet been implemented.

Arguments:

'mesh::BodyMesh' : mesh system for which to find influence coefficient matrix.

FLOWFoil.assemble_vortex_matrix — Method

assemble_vortex_matrix(mesh)

Assemble vortex coefficient matrix with full N+1 x N+1 system, including Kutta condition.

Arguments:

'mesh::BodyMesh' : Mesh System for which to solve.

FLOWFoil.camber — Method

camber(x::Float64, maxcamber::Number, camberpose::Number)

Compute camber at a given chord-normalized x-position by NACA 4-series camber equations.

FLOWFoil.defaultparameters — Method

defaultparameters(;kwargs)

Initialized parameters to defaults, but allows selective user override through keyword arguments.

Keyword Arguments:

`gamma_air::Float' : ratio of specific heats for air

`eta_crit::Float' : critical amplification factor

`eta_D::Float' : wall/wake dissipation length ratio

`GA::Float' : G - Beta locus A constant

`GB::Float' : G - Beta locus B constant

`GC::Float' : G - Beta locus C constant

`Klag::Float' : shear lag constant

`Ctau::Float' : shear stress initialization constant

`Etau::Float' : shear stree initialization exponent

`rSu::Float' : Sutherland temperature ratio

`fw::Float' : wake gap continuation factor

`dw::Float' : wake length, in airfoil chords

`epsilonw::Float' : first wake point offset, in airfoil chords

FLOWFoil.generate_mesh — Method

generatemesh(x, y; chordlength, wakelength)

Create panels from input geometry coordinates.

Arguments:

'x::Vector{Float}' : x coordinates defining airfoil geometry.

'y::Vector{Float}' : y coordinates defining airfoil geometry.

Keyword Arguments:

gaptolerance::Float' : Tolerance for how close, relative to the chord, the trailing edge nodes can be before being considered a sharp trailing edge. (default = 1e-10)

wakelength::Float' : length of wake relative to chord (default = 1)

Returns

mesh::BodyMesh : Geometry mesh, including panel nodes and trailing edge condition.

FLOWFoil.generate_mesh — Method

generate_mesh(coordinates; kwargs)

Identical to implementation with x and y separate, but here with x,y coordinates together in a single array [X Y].

Arguments:

'coordinates::Array{Float,2}' : array of both x and y coordinates (x first column, y second column).

FLOWFoil.get_a — Method

get_a(r1, d, dmag)

Calculate distance from panel to evalulation point in the panel tangent direction.

Arguments:

'r1::Vector{Float}' : vector from node1 to evalulation point.

'd::Vector{Float}' : vector from node1 to node2.

'dmag::Float' : panel length

FLOWFoil.get_d — Method

get_d(node1, node2)

Calculate panel length (between adjacent nodes).

Arguments:

'node1::Array{Float}(2)' : [x y] location of first node

'node2::Array{Float}(2)' : [x y] location of second node

Returns

'd::Vector{Float}' : vector from node1 to node2

'dmag::Float' : length of panel between node1 and node2

FLOWFoil.get_distances — Method

get_distances(node1, node2, point)

Get vectors and magnitudes for panel and between nodes and validation points.

Arguments:

'node1::Array{Float}' : [x y] position of node1.

'node2::Array{Float}' : [x y] position of node2.

'point::Array{Float}' : [x y] position of evaluation point.

Returns:

'r1::Vector{Float}' : vector from node1 to evaluation point.

'r1mag::Float' : distance from node1 to evaluation point.

'r2::Vector{Float}' : vector from node2 to evaluation point.

'r2mag::Float' : distance from node2 to evaluation point.

'd::Vector{Float}' : vector from node1 to node2.

'dmag::Float' : panel length

FLOWFoil.get_h — Method

get_h(r1, d, dmag)

Calculate distance from panel to evalulation point in the panel normal direction.

Arguments:

'r1::Vector{Float}' : vector from node1 to evalulation point.

'd::Vector{Float}' : vector from node1 to node2.

'dmag::Float' : panel length

FLOWFoil.get_inviscid_system — Method

get_inviscid_system(mesh)

Calculate, then gather the vortex and boundary condition matricies into an InviscidSystem object.

Arguments:

'mesh::BodyMesh' : BodyMesh for airfoil to analyze.

FLOWFoil.get_normal — Method

get_normal(d, dmag)

Get unit normal to panel.

Arguments:

'd::Vector{Float}' : vector from node1 to node2.

'dmag::Float' : panel length

FLOWFoil.get_orientation — Method

get_orientation(node1, node2, point)

Get angles between panel and evaluation point, ln of distances from nodes to evaluation point, and evaluation point position relative to panel.

Arguments:

'node1::Array{Float}' : [x y] position of node1.

'node2::Array{Float}' : [x y] position of node2.

'point::Array{Float}' : [x y] position of evaluation point.

Returns:

'theta1::Float' : Angle between panel and evaluation point, centered at node1.

'theta2::Float' : Angle between panel and evaluation point, centered at node2.

'ln1::Float' : Natural log of distance from node1 to evaluation point.

'ln2::Float' : Natural log of distance from node2 to evaluation point.

'h::Float' : Distance from panel to evaluation in panel normal direction.

'a::Float' : Distance from node1 to evaluation in panel tangent direction.

FLOWFoil.get_psibargamma — Method

get_psibargamma(theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\overbar{\Psi}^\gamma$

Arguments:

'theta1::Float' : angle between panel and vector from node1 to evaluation point

'theta2::Float' : angle between panel and vector from node2 to evaluation point

'ln1::Float' : value of ln(rmag1), which may be that or 0.0, depening on evaluation point location

'ln2::Float' : value of ln(rmag2), which may be that or 0.0, depening on evaluation point location

'dmag::Float' : panel length

'h::Float' : height of right triangle with hypontenuse, r1, and base, a, colinear with panel.

'a::Float' : length of base of right triangle with height, h, and hypontenuse, r1.

FLOWFoil.get_psibarsigma — Method

get_psibarsigma(theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\overbar{\Psi}^\sigma$

Arguments:

'theta1::Float' : Angle between panel and evaluation point, centered at node1.

'theta2::Float' : Angle between panel and evaluation point, centered at node2.

'ln1::Float' : Natural log of distance from node1 to evaluation point.

'ln2::Float' : Natural log of distance from node2 to evaluation point.

'h::Float' : Distance from panel to evaluation in panel normal direction.

'a::Float' : Distance from node1 to evaluation in panel tangent direction.

FLOWFoil.get_psitildegamma — Method

get_psitildegamma(psibargamma, r1mag, r2mag, theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\widetilde{\Psi}^\gamma$

Arguments:

'psibargamma::Float' : value of $\overbar{\Psi}^\gamma$

'r1mag::Float' : distance from node1 to evaluation point

'r2mag::Float' : distance from node2 to evaluation point

'theta1::Float' : angle between panel and vector from node1 to evaluation point

'theta2::Float' : angle between panel and vector from node2 to evaluation point

'ln1::Float' : value of ln(rmag1), which may be that or 0.0, depening on evaluation point location

'ln2::Float' : value of ln(rmag2), which may be that or 0.0, depening on evaluation point location

'dmag::Float' : panel length

'h::Float' : height of right triangle with hypontenuse, r1, and base, a, colinear with panel.

'a::Float' : length of base of right triangle with height, h, and hypontenuse, r1.

FLOWFoil.get_psitildesigma — Method

get_psitildesigma(psibargamma, r1mag, r2mag, theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\widetilde{\Psi}^\sigma$

Arguments:

'psibargamma::Float' : value of $\overbar{\Psi}^\gamma$

'r1mag::Float' : distance from node1 to evaluation point

'r2mag::Float' : distance from node2 to evaluation point

'theta1::Float' : angle between panel and vector from node1 to evaluation point

'theta2::Float' : angle between panel and vector from node2 to evaluation point

'dmag::Float' : panel length

'h::Float' : height of right triangle with hypontenuse, r1, and base, a, colinear with panel.

'a::Float' : length of base of right triangle with height, h, and hypontenuse, r1.

FLOWFoil.get_r — Method

function get_r(node,point)

Calculate the vector, $\mathbf{r}$, and distance, $|r|$, from the node to the evaluation point

Arguments:

node::Array{Float} : [x y] position of node

point::Array{Float} : [x y] position of point.

Returns

'r::Vector{Float}' : vector from node to evaluation point

'rmag::Float' : length of panel between node and evaluation point

FLOWFoil.get_source_influence — Method

get_source_influence(node1, node2, point)

Calculate source influence coefficients on the evaluation point from the panel between node1 and node2.

Arguments:

'node1::Array{Float}(2)' : [x y] location of node1

'node2::Array{Float}(2)' : [x y] location of node2

'point::Array{Float}(2)' : [x y] location of evaluation point

FLOWFoil.get_tangent — Method

get_tangent(d, dmag)

Get unit tangent to panel.

Arguments:

'd::Vector{Float}' : vector from node1 to node2.

'dmag::Float' : panel length

FLOWFoil.get_theta — Method

get_theta(h, a, dmag)

Get angle (in radians) between panel and vector from node2 to evaluation point.

Arguments:

'h::Float' : Distance, normal to panel, between panel and evaluation point.

'a::Float' : Distance, tangent to panel, between node1 and evaluation point.

'dmag::Float' : Panel lentgh.

FLOWFoil.get_theta — Method

get_theta(h, a)

Get angle (in radians) between panel and vector from node1 to evaluation point.

Arguments:

'h::Float' : Distance, normal to panel, between panel and evaluation point.

'a::Float' : Distance, tangent to panel, between node1 and evaluation point.

FLOWFoil.get_vortex_influence — Method

get_vortex_influence(node1, node2, point)

Calculate vortex influence coefficients on the evaluation point from the panel between node1 and node2.

Arguments:

'node1::Array{Float}(2)' : [x y] location of node1

'node2::Array{Float}(2)' : [x y] location of node2

'point::Array{Float}(2)' : [x y] location of evaluation point

FLOWFoil.initalize_viscous — Method

initalize_viscous(problem)

Initialized viscous solution (solves invscid problem, initialized wake and boundary layer, etc.)

Arguments:

problem::Problem' : Problem to solve.

Returns:

solution::ViscousSolution' : Initialized ViscousSolution

FLOWFoil.initialize_wake — Method

FLOWFoil.joukowsky — Function

joukowsky(center::Array{Float64}=[-0.1 0.1] ; N::Int=360)

Calculate the unit Joukowsky Airfoil based on a cylinder's center point.

FLOWFoil.karmantrefftz — Function

karmantrefftz(center::Array{Float64}=[-0.1 0.1], wedge::Float64=0.0;N::Int=360)

Calculate geometry of Karman-Trefftz airfoil based on the center point of a cylinder and a wedge angle.

FLOWFoil.naca4 — Function

naca4(c::Number=2.0, p::Number=4.0, t::Number=12.0, N::Int=100;bluntTE::Bool=false)

Compute x, y airfoil coordinates for N nodes, based on NACA 4-Series Parameterization.

x, y-upper, and y-lower coordinates are output. All have both leading and trailing edge points.

keyword argument: bluntTE=false (default) controls whether the blunt or sharp trailing edge formulation is used.

FLOWFoil.rawjoukowsky — Function

rawjoukowsky(center::Array{Float64}=[-0.1 0.1] ; N::Int=360)

Calculate geometry of Joukowsky Airfoil based on a cylinder's center point.

FLOWFoil.rawkarmantrefftz — Function

rawkarmanTrefftzFoil(center::Array{Float64}=[-0.1 0.1], wedge::Float64=0.0;N::Int=360)

Calculate geometry of Karman-Trefftz airfoil based on the center point of a cylinder and a wedge angle.

FLOWFoil.solve — Method

solve(problem)

Solve problem defined by the input Problem object and return the solution in a Solution object.

Arguments:

'problem::Problem' : Problem to solve

Returns:

'solution::{InviscidSolution or ViscousSolution}' : returns solution of type matching viscous flag in problem.

FLOWFoil.solve_inviscid — Method

solve_inviscid(problem)

Solves the inviscid problem.

Arguments:

'problem::Problem' : Problem to solve. viscous field must be set to false.

Returns:

'solution::InviscidSolution'

FLOWFoil.solve_inviscid_system — Method

solve_inviscid_system(inviscidsystem, mesh; debug=false)

Solve the InviscidSystem for the vortex and streamfunction strengths.

Outputs the InviscidSolution object which contains the inviscidsystem in the debug object if debug is set to true.

Arguments:

'inviscidsystem::InviscidSystem' : InviscidSystem to solve.

'mesh::BodyMesh' : BodyMesh defining geometry (to put into solution object)

Keyword Arguments:

'debug::Bool = false' : flag to indicate whether or not to output all the system details.

Returns:

'solution::InviscidSolution'

FLOWFoil.solve_viscous — Method

solve_viscous(problem)

Solves the viscous problem.

Arguments:

'problem::Problem' : Problem Definition.

Returns:

'solution::ViscousSolution'

FLOWFoil.thickness — Method

thickness(x::Float64, maxthick::Number; bluntTE::Bool=false)

Compute thickness at a given chord-normalized x-position by NACA 4-series thickness equations.