FLOWFoil

Documentation for FLOWFoil.

FLOWFoil.Freestream

FLOWFoil.Mesh

FLOWFoil.MeshSystem

FLOWFoil.Problem

FLOWFoil.Solution

FLOWFoil.assembleboundaryconditions

FLOWFoil.assemblematrixa

FLOWFoil.assemblevortexcoefficients

FLOWFoil.camber

FLOWFoil.generatemesh

FLOWFoil.get_a

FLOWFoil.get_d

FLOWFoil.get_distances

FLOWFoil.get_h

FLOWFoil.get_normal

FLOWFoil.get_orientation

FLOWFoil.get_psibargamma

FLOWFoil.get_psibarsigma

FLOWFoil.get_psitildegamma

FLOWFoil.get_psitildesigma

FLOWFoil.get_r

FLOWFoil.get_source_influence

FLOWFoil.get_tangent

FLOWFoil.get_theta

FLOWFoil.get_theta

FLOWFoil.get_vortex_influence

FLOWFoil.joukowsky

FLOWFoil.karmantrefftz

FLOWFoil.naca4

FLOWFoil.rawjoukowsky

FLOWFoil.rawkarmantrefftz

FLOWFoil.sizesystem

FLOWFoil.thickness

FLOWFoil.Freestream — Type

Freestream{TF}

Freestream Definition.

Fields:

'reynolds::Vector{Float}' : Reynolds Numbers

'density::Vector{Float}' : air density

'dynamicviscosity::Vector{Float}' : air dynamic viscosity

'mach::Vector{Float}' : Mach Numbers

'angleofattack::Vector{Float}' : Angles of attack in degrees

FLOWFoil.Mesh — Type

Mesh{TF}

Mesh for single body.

Fields:

'airfoil_nodes::Array{Array{Float,2}}' : [x y] node (panel edge) locations for airfoil

'wake_nodes::Array{Array{Float,2}}' : [x y] node (panel edge) locations for wake

'wake_midpoints::Array{Array{Float,2}}' : [x y] wake panel midpoint locations

'blunt_te::Bool' : boolean for whether or not the trailing edge is blunt or not.

Assuptions:

x and y coordinates start at the bottom trailing edge and proceed clockwise.

x and y coordinates are normalized such that the airfoil has chord length 1.0.

FLOWFoil.MeshSystem — Type

MeshSystem{TF}

System of meshes to solve.

Fields:

'meshes::Array{Mesh}' : Array of mesh objects.

'scales::Vector{Float}' : Airfoil scaling factors.

'angles::Vector{Float}' : Airfoil angles of attack.

'locations::Array{Array{TF}}' : Array of leading edge locations.

FLOWFoil.Problem — Type

Problem{TM,TS,TB}

Problem definition and method selection.

Fields:

'meshsystem::MeshSystem' : Mesh System to solve

'freestream::FreeStream' : Freestream parameters

'verbose::Bool' : Flag to print out verbose statements

'debug::Bool' : Flag to print out debugging statements

FLOWFoil.Solution — Type

Solution{TM,TF}

Output object containing solution and useful items.

Fields:

'meshsystem::MeshSystem' : Mesh System used in solution (potentially modified from input meshes).

'strengthsvec::Array{Float,2}' : singularity strengths.

'vcoeffmat::Array{Float,2}' : Vortex Coefficient Matrix used in solution.

'scoeffmat::Array{Float,2}' : Source Coefficient Matrix used in solution.

'bccoeffvec::Vector{Float}' : Boundary Coefficient Vector used in solution.

'lift::Vector{Float}' : Lift Coefficients.

'drag::Vector{Float}' : Total Drag Coefficients.

'pdrag::Vector{Float}' : Pressure Drag Coefficients.

'idrag::Vector{Float}' : Induced Drag Coefficients.

'moment::Vector{Float}' : Moment Coefficients.

'surfacevelocity::Vector{Float}' : surface velocity distribution

'surfacepressure::Vector{Float}' : surface pressure distribution

FLOWFoil.assembleboundaryconditions — Method

assembleboundaryconditions(meshsystem, freestream)

Assemble boundary condition vector.

Note that multiple operation conditions is not yet supported. If freestream fields contain more than one item, only the first will be used.

Arguments:

'meshsystem::MeshSystem' : mesh system for which to solve

'freestream::Freestream' : freestream parameters

Returns

output::type : description.

FLOWFoil.assemblematrixa — Method

assemblematrixa(meshsystem)

Assemble vortex coefficient matrix with full N+1 x N+1 system, including Kutta condition.

Arguments:

'meshsystem:MeshSystem' : Mesh System for which to solve.

FLOWFoil.assemblevortexcoefficients — Method

assemblevortexcoefficients(meshsystem)

Assemble matrix of vortex strength coefficients.

This function only assembles the NxN portion of the influence coefficient matrix. It does not include the kutta condition. Also note that multibody capabilities have not yet been implemented.

Arguments:

'meshsystem::MeshSystem' : mesh system for which to find influence coefficient matrix.

FLOWFoil.camber — Method

camber(x::Float64, maxcamber::Number, camberpose::Number)

Compute camber at a given chord-normalized x-position by NACA 4-series camber equations.

FLOWFoil.generatemesh — Method

generatemesh(x, y; chordlength, wakelength)

Create panels from input geometry coordinates.

Arguments:

'x::Vector{Float}' : x coordinates defining airfoil geometry.

'y::Vector{Float}' : y coordinates defining airfoil geometry.

Keyword Arguments:

chordlength::Float' : length of chord (default = 1)

wakelength::Float' : length of wake relative to chord (default = 1)

Returns

mesh::Mesh : Geometry mesh, including panel nodes, wake nodes, and trailing edge condition.

FLOWFoil.get_a — Method

get_a(r1, d, dmag)

Calculate distance from panel to evalulation point in the panel tangent direction.

Arguments:

'r1::Vector{Float}' : vector from node1 to evalulation point.

'd::Vector{Float}' : vector from node1 to node2.

'dmag::Float' : panel length get_a(rmag1, dmag, theta1)

FLOWFoil.get_d — Method

get_d(node1, node2)

Calculate panel length (between adjacent nodes).

Arguments:

'node1::Array{Float}(2)' : [x y] location of first node

'node2::Array{Float}(2)' : [x y] location of second node

Returns

'd::Vector{Float}' : vector from node1 to node2

'dmag::Float' : length of panel between node1 and node2

FLOWFoil.get_distances — Method

get_distances(node1, node2, point)

Get vectors and magnitudes for panel and between nodes and validation points.

Arguments:

'node1::Array{Float}' : [x y] position of node1.

'node2::Array{Float}' : [x y] position of node2.

'point::Array{Float}' : [x y] position of evaluation point.

Returns:

'r1::Vector{Float}' : vector from node1 to evaluation point.

'r1mag::Float' : distance from node1 to evaluation point.

'r2::Vector{Float}' : vector from node2 to evaluation point.

'r2mag::Float' : distance from node2 to evaluation point.

'd::Vector{Float}' : vector from node1 to node2.

'dmag::Float' : panel length

FLOWFoil.get_h — Method

get_h(r1, d, dmag)

Calculate distance from panel to evalulation point in the panel normal direction.

Arguments:

'r1::Vector{Float}' : vector from node1 to evalulation point.

'd::Vector{Float}' : vector from node1 to node2.

'dmag::Float' : panel length

FLOWFoil.get_normal — Method

get_normal(d, dmag)

Get unit normal to panel.

Arguments:

'd::Vector{Float}' : vector from node1 to node2.

'dmag::Float' : panel length

FLOWFoil.get_orientation — Method

get_orientation(node1, node2, point)

Get angles between panel and evaluation point, ln of distances from nodes to evaluation point, and evaluation point position relative to panel.

Arguments:

'node1::Array{Float}' : [x y] position of node1.

'node2::Array{Float}' : [x y] position of node2.

'point::Array{Float}' : [x y] position of evaluation point.

Returns:

'theta1::Float' : Angle between panel and evaluation point, centered at node1.

'theta2::Float' : Angle between panel and evaluation point, centered at node2.

'ln1::Float' : Natural log of distance from node1 to evaluation point.

'ln2::Float' : Natural log of distance from node2 to evaluation point.

'h::Float' : Distance from panel to evaluation in panel normal direction.

'a::Float' : Distance from node1 to evaluation in panel tangent direction.

FLOWFoil.get_psibargamma — Method

get_psibargamma(theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\overbar{\Psi}^\gamma$

Arguments:

'theta1::Float' : angle between panel and vector from node1 to evaluation point

'theta2::Float' : angle between panel and vector from node2 to evaluation point

'ln1::Float' : value of ln(rmag1), which may be that or 0.0, depening on evaluation point location

'ln2::Float' : value of ln(rmag2), which may be that or 0.0, depening on evaluation point location

'dmag::Float' : panel length

'h::Float' : height of right triangle with hypontenuse, r1, and base, a, colinear with panel.

'a::Float' : length of base of right triangle with height, h, and hypontenuse, r1.

FLOWFoil.get_psibarsigma — Method

get_psibarsigma(theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\overbar{\Psi}^\sigma$

Arguments:

'theta1::Float' : Angle between panel and evaluation point, centered at node1.

'theta2::Float' : Angle between panel and evaluation point, centered at node2.

'ln1::Float' : Natural log of distance from node1 to evaluation point.

'ln2::Float' : Natural log of distance from node2 to evaluation point.

'h::Float' : Distance from panel to evaluation in panel normal direction.

'a::Float' : Distance from node1 to evaluation in panel tangent direction.

FLOWFoil.get_psitildegamma — Method

get_psitildegamma(psibargamma, r1mag, r2mag, theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\widetilde{\Psi}^\gamma$

Arguments:

'psibargamma::Float' : value of $\overbar{\Psi}^\gamma$

'r1mag::Float' : distance from node1 to evaluation point

'r2mag::Float' : distance from node2 to evaluation point

'theta1::Float' : angle between panel and vector from node1 to evaluation point

'theta2::Float' : angle between panel and vector from node2 to evaluation point

'ln1::Float' : value of ln(rmag1), which may be that or 0.0, depening on evaluation point location

'ln2::Float' : value of ln(rmag2), which may be that or 0.0, depening on evaluation point location

'dmag::Float' : panel length

'h::Float' : height of right triangle with hypontenuse, r1, and base, a, colinear with panel.

'a::Float' : length of base of right triangle with height, h, and hypontenuse, r1.

FLOWFoil.get_psitildesigma — Method

get_psitildesigma(psibargamma, r1mag, r2mag, theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\widetilde{\Psi}^\sigma$

Arguments:

'psibargamma::Float' : value of $\overbar{\Psi}^\gamma$

'r1mag::Float' : distance from node1 to evaluation point

'r2mag::Float' : distance from node2 to evaluation point

'theta1::Float' : angle between panel and vector from node1 to evaluation point

'theta2::Float' : angle between panel and vector from node2 to evaluation point

'dmag::Float' : panel length

'h::Float' : height of right triangle with hypontenuse, r1, and base, a, colinear with panel.

'a::Float' : length of base of right triangle with height, h, and hypontenuse, r1.

FLOWFoil.get_r — Method

function get_r(node,point)

Calculate the vector, $\mathbf{r}$, and distance, $|r|$, from the node to the evaluation point

Arguments:

node::Array{Float} : [x y] position of node

point::Array{Float} : [x y] position of point.

Returns

'r::Vector{Float}' : vector from node to evaluation point

'rmag::Float' : length of panel between node and evaluation point

FLOWFoil.get_source_influence — Method

get_source_influence(node1, node2, point)

Calculate source influence coefficients on the evaluation point from the panel between node1 and node2.

Arguments:

'node1::Array{Float}(2)' : [x y] location of node1

'node2::Array{Float}(2)' : [x y] location of node2

'point::Array{Float}(2)' : [x y] location of evaluation point

FLOWFoil.get_tangent — Method

get_tangent(d, dmag)

Get unit tangent to panel.

Arguments:

'd::Vector{Float}' : vector from node1 to node2.

'dmag::Float' : panel length

FLOWFoil.get_theta — Method

get_theta(h, a, dmag)

Get angle (in radians) between panel and vector from node to evaluation point.

Arguments:

'h::Float' : Distance, normal to panel, between panel and evaluation point.

'a::Float' : Distance, tangent to panel, between node1 and evaluation point.

'dmag::Float' : Panel lentgh.

FLOWFoil.get_theta — Method

get_theta(h, a)

Get angle (in radians) between panel and vector from node to evaluation point.

Arguments:

'h::Float' : Distance, normal to panel, between panel and evaluation point.

'a::Float' : Distance, tangent to panel, between node1 and evaluation point.

FLOWFoil.get_vortex_influence — Method

get_vortex_influence(node1, node2, point)

Calculate vortex influence coefficients on the evaluation point from the panel between node1 and node2.

Arguments:

'node1::Array{Float}(2)' : [x y] location of node1

'node2::Array{Float}(2)' : [x y] location of node2

'point::Array{Float}(2)' : [x y] location of evaluation point

FLOWFoil.joukowsky — Function

joukowsky(center::Array{Float64}=[-0.1 0.1] ; N::Int=360)

Calculate the unit Joukowsky Airfoil based on a cylinder's center point.

FLOWFoil.karmantrefftz — Function

karmantrefftz(center::Array{Float64}=[-0.1 0.1], wedge::Float64=0.0;N::Int=360)

Calculate geometry of Karman-Trefftz airfoil based on the center point of a cylinder and a wedge angle.

FLOWFoil.naca4 — Function

naca4(c::Number=2.0, p::Number=4.0, t::Number=12.0, N::Int=100;bluntTE::Bool=false)

Compute x, y airfoil coordinates for N nodes, based on NACA 4-Series Parameterization.

x, y-upper, and y-lower coordinates are output. All have both leading and trailing edge points.

keyword argument: bluntTE=false (default) controls whether the blunt or sharp trailing edge formulation is used.

FLOWFoil.rawjoukowsky — Function

rawjoukowsky(center::Array{Float64}=[-0.1 0.1] ; N::Int=360)

Calculate geometry of Joukowsky Airfoil based on a cylinder's center point.

FLOWFoil.rawkarmantrefftz — Function

rawkarmanTrefftzFoil(center::Array{Float64}=[-0.1 0.1], wedge::Float64=0.0;N::Int=360)

Calculate geometry of Karman-Trefftz airfoil based on the center point of a cylinder and a wedge angle.

FLOWFoil.sizesystem — Method

sizesystem(meshsystem)

Count size of inviscid system matrix.

Arguments:

'meshsystem::MeshSystem' : The system for which to calculate the linear system size.

FLOWFoil.thickness — Method

thickness(x::Float64, maxthick::Number; bluntTE::Bool=false)

Compute thickness at a given chord-normalized x-position by NACA 4-series thickness equations.