Public

FLOWFoil.AirfoilTools.BasicBSpline

FLOWFoil.AirfoilTools.CST

FLOWFoil.AirfoilTools.Joukowsky

FLOWFoil.AirfoilTools.KarmanTrefftz

FLOWFoil.AirfoilTools.KarmanTrefftz

FLOWFoil.AirfoilTools.ModifiedPARSEC

FLOWFoil.AirfoilTools.NACA4

FLOWFoil.AirfoilTools.PARSEC

FLOWFoil.AirfoilTools.basic_bspline

FLOWFoil.AirfoilTools.bernstein

FLOWFoil.AirfoilTools.calculate_modified_parsec_coefficients

FLOWFoil.AirfoilTools.calculate_parsec_coefficients

FLOWFoil.AirfoilTools.computed_camber_line

FLOWFoil.AirfoilTools.cst

FLOWFoil.AirfoilTools.definespline

FLOWFoil.AirfoilTools.determine_cst

FLOWFoil.AirfoilTools.determine_modified_parsec

FLOWFoil.AirfoilTools.determine_naca4

FLOWFoil.AirfoilTools.determine_parsec

FLOWFoil.AirfoilTools.dot

FLOWFoil.AirfoilTools.flip!

FLOWFoil.AirfoilTools.get_spline_coordinates

FLOWFoil.AirfoilTools.half_cst

FLOWFoil.AirfoilTools.joukowsky

FLOWFoil.AirfoilTools.joukowsky_flow

FLOWFoil.AirfoilTools.karman_trefftz

FLOWFoil.AirfoilTools.modified_parsec

FLOWFoil.AirfoilTools.naca4

FLOWFoil.AirfoilTools.naca4_camber

FLOWFoil.AirfoilTools.naca4_thickness

FLOWFoil.AirfoilTools.naca65_scaled

FLOWFoil.AirfoilTools.norm

FLOWFoil.AirfoilTools.normalize_coordinates!

FLOWFoil.AirfoilTools.parsec

FLOWFoil.AirfoilTools.position_coordinates!

FLOWFoil.AirfoilTools.refine_trailing_edge

FLOWFoil.AirfoilTools.repanel_airfoil

FLOWFoil.AirfoilTools.rotate_coordinates!

FLOWFoil.AirfoilTools.scaled_camber

FLOWFoil.AirfoilTools.scaled_slope

FLOWFoil.AirfoilTools.scaled_thickness

FLOWFoil.AirfoilTools.slopes65

FLOWFoil.AirfoilTools.split_cosine_spacing

FLOWFoil.AirfoilTools.split_upper_lower

FLOWFoil.AirfoilTools.thickness65

FLOWFoil.AirfoilTools.whole_cosine_spacing

FLOWFoil.AirfoilTools.z_from_parsec_coefficients

FLOWFoil.AirfoilTools.zero_z_te!

Private

FLOWFoil.AirfoilTools.bernstein — Method

bernstein(r, n, x)

Bernstein Basis Function: binomial(n, r) .* x .^ r .* (1 .- x) .^ (n .- r)

FLOWFoil.AirfoilTools.calculate_modified_parsec_coefficients — Function

calculate_modified_parsec_coefficients(p, uppper_side)

Calculate the PARSEC coefficients using modified parameters (see parsec() docstring) for either the top or bottom curve.

Arguments:

p::NamedTuple : Named tuple of ModifiedPARSEC paramters including:

leading_edge_radius::Float : Leading edge radius

maximum_thickness_x::Float : chordwise position of maximum thickness

maximum_thikcness_z::Float : z-coordinate at maximum thickness

curvature::Float : second derivative of surface geometry at maximum thickness

trailing_edge_tangent::Float : trailing edge tangent angle, radians

trailing_edge_z::Float : z-position of trailing edge

side::Number : +1 for upper side, -1 for lower side

FLOWFoil.AirfoilTools.calculate_parsec_coefficients — Function

calculate_parsec_coefficients(p, side=1)

Calculate the PARSEC coefficients using standard parameters (see parsec() docstring) for either the top or bottom curve.

Arguments:

p::NamedTuple : NamedTuple of PARSEC standard paramters including:

[1]: leading_edge_radius : Leading edge radius

[2]: X : chordwise position of maximum thickness

[3]: Z : z-coordinate at maximum thickness

[4]: curvature : second derivative of surface geometry at maximum thickness

[5]: trailing_edge_angle : trailing edge angle

[6]: boattail_angle : boat-tail angle

[7]: trailing_edge_gap : z-distance between upper and lower surface trailing edge points

[8]: trailing_edge_z : z-position of center of trailing edge

FLOWFoil.AirfoilTools.computed_camber_line — Method

computed_camber_line(
a,
cli,
x

)

This computes the corresponding camber for a given x value for the NACA 65 series.

Arguments:

a::TF : Mean-line designation, fraction of chord from leading edge over which design load is uniform. Should be listed in the airfoil descirption

cli::TF : Design lift coefficient in tenths of chord. Usually first number after the 2nd dash (ie NACA 65-3-818 would input 0.8 for cli)

x::TF : x value (note that x must be normalized by the chord length)

Returns:

yc::Float : Amount of camber y_c for a given x value on the airfoil

FLOWFoil.AirfoilTools.definespline — Method

definespline()

Calculate the x and y location of the control points, and weight them according to the weighting vector. Also, provide generic knot vectors.

Inputs:

thirdctrlptposition : x position of the 2nd and 6th control points in degree 3.

leadingedgeradius : Leading edge radius

wedge_angle : The trailing edge wedge angle

trailingedgegap : The trailing edge gap

weights : vector weights for the control points, leave empty for unity.

Keyword Arguments:

Outputs:

knots : the knot vector of the curve

controlpoints : control point vector (x,y,w)

FLOWFoil.AirfoilTools.dot — Method

dot(A, B) = sum(a * b for (a, b) in zip(A, B))

A faster dot product.

FLOWFoil.AirfoilTools.get_spline_coordinates — Method

    get_spline_coordinates()

Get plotable coordinates from the weighted control points.

Arguments:

knots::AbstractArray{Float} : knot vector

controlpoints::AbstractArray{Float,2}: Control Point matrix

degree::Int : degree of the NURBS curve

Keyword Arguments:

N::Int : number of coordinates, or panels that will be generated

Returns:

x::AbstractArray{Float} : x Airfoil coordinates

z::AbstractArray{Float} : z Airfoil coordinates

FLOWFoil.AirfoilTools.half_cst — Function

half_cst(coefficients, x=cosine_spacing(N), dz=0.0, N1=0.5, N2=1.0)

Define upper or lower side of airfoil using CST parameterization.

Arguments:

coefficients::AbstractArray{Float} : Vector of CST coefficients.

Keyword Arguments:

dz::Float=0.0 : trailing edge gap

N1::Float=0.5 : Class shape parameter 1

N2::Float=1.0 : Class shape parameter 2

FLOWFoil.AirfoilTools.joukowsky_flow — Method

joukowsky_flow(center, radius, alpha, vinf; N=361)

Calculate the analytic surface velocities and pressures as well as lift coefficient for a joukowsky airfoil.

Arguments:

center::AbstractArray{Float} : [x z] location of circle center relative to origin

radius::Float : Radius of circle

alpha::Float : Angle of attack in degrees

vinf::Float : Freestream velocity

Keyword Arguments:

N::Int : Total number of coordinates to use. Can be even or odd, but it is recommended to be odd for a clear leading edge point.

Returns:

vsurf::AbstractArray{Float} : Magnitude of surface velocities at the nodes

cpsurf::AbstractArray{Float} : Surface pressures at the nodes

cl::Float : Lift coefficient

FLOWFoil.AirfoilTools.naca4_camber — Method

naca4_camber(x, max_camber, max_camber_pos)

Compute camber at a given chord-normalized x-position by NACA 4-series camber equations.

Arguments:

x::Float : x position along chordline

max_camber::Float64 : Maximum camber value

max_camber_pos::Float64 : Position of maximum camber

FLOWFoil.AirfoilTools.naca4_thickness — Method

naca4_thickness(x, maxthick; blunt_te=false)

Compute thickness at a given chord-normalized x-position by NACA 4-series thickness equations.

Arguments:

x::Float : x position along chordlin, markersize=3, markershape=:squaree

maxthick::Float : Maximum thickness value

Keyword Arguments:

blunt_te::Bool : Flag whether trailing edge is blunt or not

FLOWFoil.AirfoilTools.naca65_scaled — Method

Assumes x in non-dimensional range [0.0,1.0]

Description from NACA Research Memorandum L51G31: "Systematic Two-dimensional Cascade Tests of NACA 65-Series Compressor Blades at Low Speeds:"

The 65-series compressor blade family is formed by combining a basic thickness form with cambered mean lines. The basic thickness form used is the NACA 65(216)-010 thickness form with the ordinates increased by 0.0015 times the chordwise stations to provide slightly, increased thickness toward the trailing edge. In the scaled case, it was not derived for 10-percent thickness but was scaled down from the NACA 65,2-016 airfoil. The scaling procedure gives the best results whep it is restricted to maximum thickness changes of a few percent. The NACA 65-010 basic thickness has also been derived. These thickness forms differ slightly but are considered to be interchangeable.

The basic mean line used is the a=1.0 mean line. The amount of camber is for the design lift coefficient for the isolated airfoil with clo of 1.0. Both ordinates and slopes are scaled directly to obtain other cambers. Cambered blade sections are obtained by applying the thickness perpendicular to the mean line at stations laid out along the chord line. In the designation the camber is given by the first number after the dash in tenths of clo. For example, the NACA 65-810 and NACA 65-(12)10 blade sections are cambered for clo = 0.8 and clo = 1.2, respectively.

This specific function plots the NACA 65-010 series which uses special mean line and thickness form coordinates. It scales those coordinates based on the cl_o value.

Arguments:

clo::TF : Design lift coefficient in tenths of chord. Usually first number after the 2nd dash (ie NACA 65-3-818 would input 0.8 for clo)

Keyword Arguments:

x::Vector{TF} = nothing : input x values if specificing x values manually

split::Boolean = false : if true, then the output will be split between top and bottom coordinates

extra_blending::Boolean = false : If desired number of points is large (> 300ish) then set to true and it will add some extra blending if desired. Note: This is generally not needed!

ouptuts:

x::Vector{TF} : x coordinates

y::Vector{TF} : y coordinates

FLOWFoil.AirfoilTools.norm — Method

norm(A) = sqrt(mapreduce(x -> x^2, +, A))

A faster 2-norm.

FLOWFoil.AirfoilTools.scaled_camber — Method

From NACA Research Memorandum L51G31: "Systematic Two-dimensional Cascade Tests of NACA 65-Series Compressor Blades at Low Speeds" Table 2

FLOWFoil.AirfoilTools.scaled_slope — Method

From NACA Research Memorandum L51G31: "Systematic Two-dimensional Cascade Tests of NACA 65-Series Compressor Blades at Low Speeds" Table 2

FLOWFoil.AirfoilTools.scaled_thickness — Method

From NACA Research Memorandum L51G31: "Systematic Two-dimensional Cascade Tests of NACA 65-Series Compressor Blades at Low Speeds" Table 1

FLOWFoil.AirfoilTools.slopes65 — Method

slopes65(
x,
clo,
a;
slope_tolerance = 0.000001

)

This computes the corresponding slopes of the mean line for a given x value on the airfoil

Arguments:

x::TF : x value (note that x must be normalized by the chord length)

clo::TF : Design lift coefficient in tenths of chord. Usually first number after the 2nd dash (ie NACA 65-3-818 would input 0.8 for cli)

a::TF : Mean-line designation, fraction of chord from leading edge over which design load is uniform. Should be listed in the airfoil descirption

Keyword Arguments:

slope_tolerance::Float = 0.000001 : step size taken for the derivative

Returns:

tan_theta::Float : Tangent of the mean line slope

sin_theta::Float : Sine of the mean line slope

cos_theta::Float : Cosine of the mean line slope

FLOWFoil.AirfoilTools.thickness65 — Method

thickness65(
series_number,
xpt #value of x that the desired y_t is

)

This computes the thicknes coordinate and leading edge radius from tabulated thickness forms

Arguments:

series_number::String : Airfoil series number (see naca65 doc string for more details)

xpt::TF : Desired x value to compute the thickness value

Returns:

y_t::Float : thickness value at specified x value

leading_edge_radius::Float : Tabulated leading edge thickness - make sure it is in x/c units!

FLOWFoil.AirfoilTools.z_from_parsec_coefficients — Function

z_from_parsec_coefficients(a, N::Int=80)

Calculate the x,z airfoil coordinates using the PARSEC polynomial.

Arguments:

a::AbstractArray{Float} : the PARSEC coefficients.