Common Parameterizations

Included in the AirfoilTools module are the following common airfoil parameterizations for convenience in generating various airfoil geometries.

Parameter Types

Each of the parameterizations below have associated composite types whose fields are the parameters used in defining the airfoil geometries. Each of these composite types is defined using the @kwdef macro such that the user does not need to remember the order of the fields, but can simply define the fields as though they were keyword arguments. In general, few of the fields are given defaults with the exception of thigns like trailing edge gap or z-positions, which are always defaulted to zero. In addition, some parameterization methods have specific values inherent to their methods. These are exposed to the user for convenience, but are also defaulted to the values inherent in the parameterization method.

Coordinates

For all of the parameterizations, the x-coordinates are defined to be along the chord-wise direction, and the z-coordinates are orthogonal to the x-coordinates. This is done to be similar to the standard airframe coordinate system, with the x-direction being positve toward the back of the airframe, and the z-direction being up.

In addition all coordinates are given clockwise from the trailing edge, even if the coordinates are split between upper and lower sides. In other words, the coordinates are given from the lower trailing edge to the leading edge, then the leading edge back to the upper side trailing edge. In general, if the coordinates are given in an upper and lower split, the leading edge point is repeated.

NACA 4-series

We begin with the standard NACA 4-series airfoil, defining the parameters in the NACA4 type as

FLOWFoil.AirfoilTools.NACA4 — Type

Fields:

max_camber::Float=2.0 : maximum camber in % chord

max_camber_pos::Float=4.0 : x-position of maximum camber point in 1/10 chord

max_thickness::Float=12.0 : maximum thickness in % chord

blunt_te::Bool=false : flag for whether to use the blunt trailing edge NACA 4-series definition or not.

Defaults

The NACA4 parameter type is the only one with fully defined defaults, which happen to default to the NACA 2412 airfoil with a sharp trailing edge due to its ubiquity.

A NACA 4-series airfoil can be defined with either of the naca4 methods.

FLOWFoil.AirfoilTools.naca4 — Function

naca4(parameters::NACA4; N=161, x=nothing, split=false)

Compute x, z airfoil coordinates for N nodes, based on NACA 4-Series Parameterization.

Arguments:

parameters::NACA4 : NACA 4-series parameters

Keyword Arguments:

N::Int : Total number of coordinates to use. This values should be odd, but if not, the number of points returned will be N-1.

x::AbstractArray{Float} : x coordinates (cosine spaced coordinates used by default)

split::Bool : Flag wheter to split into upper and lower halves.

Returns:

If split == false:

x::AbstractArray{Float} : Vector of x coordinates, clockwise from trailing edge.

z::AbstractArray{Float} : Vector of z coordinates, clockwise from trailing edge.

If split == true:

xl::AbstractArray{Float} : Vector of lower half of x coordinates from trailing edge to leading edge.

xu::AbstractArray{Float} : Vector of upper half of x coordinates from leading edge to trailing edge.

zl::AbstractArray{Float} : Vector of lower half of z coordinates from trailing edge to leading edge.

zu::AbstractArray{Float} : Vector of upper half of z coordinates from leading edge to trailing edge.

naca4(c=2.0, p=4.0, t=12.0; N=161, x=nothing, blunt_te=false, split=false)

Compute x, z airfoil coordinates for N nodes, based on NACA 4-Series Parameterization.

Arguments:

c::Float : Maximum camber value (percent of chord)

p::Float : Position along chord (in 10ths of chord) where maximum naca4_camber lies

t::Float : Maximum thickness of airfoil in percent chord

Keyword Arguments:

N::Int : Total number of coordinates to use. This values should be odd, but if not, the number of points returned will be N-1.

x::AbstractArray{Float} : x-coordinates (cosine spaced coordinates used by default)

blunt_te::Bool : Flag whether trailing edge is blunt or not

split::Bool : Flag wheter to split into upper and lower halves.

Returns:

If split == false:

x::AbstractArray{Float} : Vector of x coordinates, clockwise from trailing edge.

z::AbstractArray{Float} : Vector of z coordinates, clockwise from trailing edge.

If split == true:

xl::AbstractArray{Float} : Vector of lower half of x coordinates from trailing edge to leading edge.

xu::AbstractArray{Float} : Vector of upper half of x coordinates from leading edge to trailing edge.

zl::AbstractArray{Float} : Vector of lower half of z coordinates from trailing edge to leading edge.

zu::AbstractArray{Float} : Vector of upper half of z coordinates from leading edge to trailing edge.

For a given set of x,z coordinates a best-fit NACA 4-series airfoil can be found using

FLOWFoil.AirfoilTools.determine_naca4 — Function

determine_naca4(x,z)

Calculate NACA 4-series parameters based on input x,z coordinates.

Returns:

parameters::NACA4 : a parameter object of type NACA4.

Conformal Mapping

AirfoilTools includes two conformal mapping methods; the first is the Joukowsky method, and the second is the KarmanTrefftz method (which is a variation on Joukowsky allowing for non-cusped trailing edges)

Joukowsky

The Joukowsky parameter type is defined as

FLOWFoil.AirfoilTools.Joukowsky — Type

Fields:

center::AbstractArray{Float} : [x z] location of center of circle relative to origin

radius::Float : radius of circle

A Joukowsky airfoil can be generated with

FLOWFoil.AirfoilTools.joukowsky — Function

joukowsky(parameters::Joukowsky; N, fortest=false, normalize=true, split=false)

Joukowsky airfoil parameterization.

Arguments:

parameters::Joukowsky : Joukowsky parameters

Keyword Arguments:

N::Int : Total number of coordinates to use. Can be even or odd, but it is recommended to be odd for a clear leading edge point.

fortest::Bool : Flag to output non-coordinate paramters used in 'joukowsky_flow()'

normalize::Bool : Flag whether to normalize to unit chord and translate the leading edge to zero.

split::Bool : Flag wheter to split output into upper and lower surfaces.

Returns:

IF split == False:

x::AbstractArray{Float} : Array of x coordinates

z::AbstractArray{Float} : Array of z coordinates

IF split == True:

xu::AbstractArray{Float} : Array of upper half of x coordinates

xl::AbstractArray{Float} : Array of lower half of x coordinates

zu::AbstractArray{Float} : Array of upper half of z coordinates

zl::AbstractArray{Float} : Array of lower half of z coordinates

joukowsky(center, radius; N, fortest=false, normalize=true, split=false)

Joukowsky airfoil parameterization.

Arguments:

center::AbstractArray{Float} : [x z] location of center of circle relative to origin

radius::Float : radius of circle

Keyword Arguments:

N::Int : Total number of coordinates to use. Can be even or odd, but it is recommended to be odd for a clear leading edge point.

fortest::Bool : Flag to output non-coordinate paramters used in 'joukowsky_flow()'

normalize::Bool : Flag whether to normalize to unit chord and translate the leading edge to zero.

split::Bool : Flag wheter to split output into upper and lower surfaces.

Returns:

IF split == False:

x::AbstractArray{Float} : Array of x coordinates

z::AbstractArray{Float} : Array of z coordinates

IF split == True:

xu::AbstractArray{Float} : Array of upper half of x coordinates

xl::AbstractArray{Float} : Array of lower half of x coordinates

zu::AbstractArray{Float} : Array of upper half of z coordinates

zl::AbstractArray{Float} : Array of lower half of z coordinates

Karman-Trefftz

The KarmanTrefftz parameter type is defined as

FLOWFoil.AirfoilTools.KarmanTrefftz — Type

Fields:

beta::Float : angle, in radians indicating center of circle relative to origin

radius::Float : radius of circle

wedge_angle::Float : angle, in radians, of airfoil wedge angle

FLOWFoil.AirfoilTools.KarmanTrefftz — Method

KarmanTrefftz(center, wedge_angle)

Arguments

center::AbstractArray{Float} : [x z] location of circle center relative to origin

wedge_angle::Float : angle, in radians, of airfoil wedge angle

Returns

parameters::KarmanTrefftz : KarmanTrefftz parameters

A Karman-Trefftz airfoil can be generated with

FLOWFoil.AirfoilTools.karman_trefftz — Function

karman_trefftz(paramters::KarmanTrefftz; N=361, normalize=true, split=false)

Karman-Trefftz airfoil parameterization based on angle beta, raidus, and wedge angle.

Arguments:

paramters::KarmanTrefftz : KarmanTrefftz parameters

Keyword Arguments:

N::Int : Total number of coordinates to use. Can be even or odd, but it is recommended to be odd for a clear leading edge point.

normalize::Bool : Flag whether to normalize output to unit chord and shift to have leading edge at zero.

split::Bool : Flag wheter to split into upper and lower halves.

Returns:

If split == false:

x::AbstractArray{Float} : Vector of x coordinates, clockwise from trailing edge.

z::AbstractArray{Float} : Vector of z coordinates, clockwise from trailing edge.

If split == true:

xl::AbstractArray{Float} : Vector of lower half of x coordinates from trailing edge to leading edge.

xu::AbstractArray{Float} : Vector of upper half of x coordinates from leading edge to trailing edge.

zl::AbstractArray{Float} : Vector of lower half of z coordinates from trailing edge to leading edge.

zu::AbstractArray{Float} : Vector of upper half of z coordinates from leading edge to trailing edge.

karman_trefftz(beta, radius, wedge_angle; N=361, normalize=true, split=false)

Karman-Trefftz airfoil parameterization based on angle beta, raidus, and wedge angle.

Arguments:

beta::Float : angle, in radians indicating center of circle relative to origin

radius::Float : radius of circle

wedge_angle::Float : angle, in radians, of airfoil wedge angle

Keyword Arguments:

N::Int : Total number of coordinates to use. Can be even or odd, but it is recommended to be odd for a clear leading edge point.

normalize::Bool : Flag whether to normalize output to unit chord and shift to have leading edge at zero.

split::Bool : Flag wheter to split into upper and lower halves.

Returns:

If split == false:

x::AbstractArray{Float} : Vector of x coordinates, clockwise from trailing edge.

z::AbstractArray{Float} : Vector of z coordinates, clockwise from trailing edge.

If split == true:

xl::AbstractArray{Float} : Vector of lower half of x coordinates from trailing edge to leading edge.

xu::AbstractArray{Float} : Vector of upper half of x coordinates from leading edge to trailing edge.

zl::AbstractArray{Float} : Vector of lower half of z coordinates from trailing edge to leading edge.

zu::AbstractArray{Float} : Vector of upper half of z coordinates from leading edge to trailing edge.

karman_trefftz(center, wedge_angle; N=361, normalize=true, split=false)

Identical to karman_trefftz(beta, radius, wedge_angle) but using center-based version.

Arguments:

center::AbstractArray{Float} : [x z] location of circle center relative to origin

wedge_angle::Float : angle, in radians, of airfoil wedge angle

Keyword Arguments:

N::Int : Total number of coordinates to use. Can be even or odd, but it is recommended to be odd for a clear leading edge point.

normalize::Bool : Flag whether to normalize out put to unit chord and shift to have leading edge at zero.

split::Bool : Flag wheter to split into upper and lower halves.

Returns:

IF split == False:

x::AbstractArray{Float} : Array of x coordinates

z::AbstractArray{Float} : Array of z coordinates

IF split == True:

xu::AbstractArray{Float} : Array of upper half of x coordinates

xl::AbstractArray{Float} : Array of lower half of x coordinates

zu::AbstractArray{Float} : Array of upper half of z coordinates

zl::AbstractArray{Float} : Array of lower half of z coordinates

Class Shape Transformations (CST)

The CST parameter type is defined as

FLOWFoil.AirfoilTools.CST — Type

Fields:

upper_coefficients::AbstractArray{Float} : Vector of coefficients defining the upper side.

lower_coefficients::AbstractArray{Float} : Vector of coefficients defining the lower side.

trailing_edge_zu::Float=0.0 : z-position of the upper side trailing edge.

trailing_edge_zl::Float=0.0 : z-position of the lower side trailing edge.

N1::Float=0.5 : inherent parameter for round-nosed airfoils.

N2::Float=1.0 : inherent parameter for sharp trailing edge (with optional blunt trailing edge) airfoils.

A CST airfoil can be generated with

FLOWFoil.AirfoilTools.cst — Function

cst(
    CST,
    N::Integer=80,
    x=split_cosine_spacing(N),
    split=false,
)

Obtain airfoil coordiantes (clockwise from trailing edge) from the class shape transformation (CST) parameterization.

Arguments:

parameters::CST : CST parameters for airfoil.

Keyword Arguments:

N::Integer=80 : number of points to use for each side

x::AbstractArray{Float}=split_cosine_spacing(N) : x-coordinates to use.

trailing_edge_zu::Float=0.0 : upper side trailing edge gap

trailing_edge_zl::Float=0.0 : lower side trailing edge gap

N1::Float=0.5 : Class shape parameter 1

N2::Float=1.0 : Class shape parameter 2

split::Bool=false : if true, returns upper and lower coordinates separately as xl, xu, zl, zu rather than just x, z.

Returns:

If split == false:

x::AbstractArray{Float} : Vector of x coordinates, clockwise from trailing edge.

z::AbstractArray{Float} : Vector of z coordinates, clockwise from trailing edge.

If split == true:

xl::AbstractArray{Float} : Vector of lower half of x coordinates from trailing edge to leading edge.

xu::AbstractArray{Float} : Vector of upper half of x coordinates from leading edge to trailing edge.

zl::AbstractArray{Float} : Vector of lower half of z coordinates from trailing edge to leading edge.

zu::AbstractArray{Float} : Vector of upper half of z coordinates from leading edge to trailing edge.

cst(
    upper_coefficients,
    lower_coefficients;
    N::Integer=80,
    x=split_cosine_spacing(N),
    trailing_edge_zu=0.0,
    trailing_edge_zl=0.0,
    N1=0.5,
    N2=1.0,
    split=false,
)

Obtain airfoil coordiantes (clockwise from trailing edge) from the class shape transformation (CST) parameterization.

Arguments:

upper_coefficients::AbstractArray{Float} : Vector of CST coefficients for upper side of airfoil.

lower_coefficients::AbstractArray{Float} : Vector of CST coefficients for lower side of airfoil.

Keyword Arguments:

N::Integer=80 : number of points to use for each side

x::AbstractArray{Float}=split_cosine_spacing(N) : x-coordinates to use.

trailing_edge_zu::Float=0.0 : upper side trailing edge gap

trailing_edge_zl::Float=0.0 : lower side trailing edge gap

N1::Float=0.5 : Class shape parameter 1

N2::Float=1.0 : Class shape parameter 2

split::Bool=false : if true, returns upper and lower coordinates separately as xl, xu, zl, zu rather than just x, z.

Returns:

If split == false:

x::AbstractArray{Float} : Vector of x coordinates, clockwise from trailing edge.

z::AbstractArray{Float} : Vector of z coordinates, clockwise from trailing edge.

If split == true:

xl::AbstractArray{Float} : Vector of lower half of x coordinates from trailing edge to leading edge.

xu::AbstractArray{Float} : Vector of upper half of x coordinates from leading edge to trailing edge.

zl::AbstractArray{Float} : Vector of lower half of z coordinates from trailing edge to leading edge.

zu::AbstractArray{Float} : Vector of upper half of z coordinates from leading edge to trailing edge.

For a given set of x,z coordinates a best-fit CST airfoil can be found using

FLOWFoil.AirfoilTools.determine_cst — Function

determine_cst(
    x,
    z;
    n_upper_coefficients::Integer=8,
    n_lower_coefficients::Integer=8,
    trailing_edge_zu=0.0,
    trailing_edge_zl=0.0,
    N1=0.5,
    N2=1.0,
)

Determine best-fit CST parameters using a least squares solve.

Arguments:

x::AbstractArray{Float} : vector of x-coordinates.

z::AbstractArray{Float} : vector of z-coordinates.

Keyword Arguments:

n_upper_coefficients::Integer=8 : number of upper side coefficients to fit

n_lower_coefficients::Integer=8 : number of lower side coefficients to fit

N1::Float=0.5 : Class shape parameter 1

N2::Float=1.0 : Class shape parameter 2

Returns:

paramters::CST : CST paramters for airfoil.

determine_cst(
    xl,
    xu,
    zl,
    zu;
    n_upper_coefficients::Integer=8,
    n_lower_coefficients::Integer=8,
    N1=0.5,
    N2=1.0,
)

Determine best-fit CST parameters for upper and lower sides of airfoil using a least squares solve.

Arguments:

xl::AbstractArray{Float} : vector of lower side x-coordinates.

xu::AbstractArray{Float} : vector of upper side x-coordinates.

zl::AbstractArray{Float} : vector of lower side z-coordinates.

zu::AbstractArray{Float} : vector of upper side z-coordinates.

Keyword Arguments:

n_upper_coefficients::Integer=8 : number of upper side coefficients to fit

n_lower_coefficients::Integer=8 : number of lower side coefficients to fit

N1::Float=0.5 : Class shape parameter 1

N2::Float=1.0 : Class shape parameter 2

Returns:

paramters::CST : CST paramters for airfoil.

Basic B-Spline

The basic B-Spline parameterization comes from "Universal Airfoil Parametrization Using B-Splines" by Rajnarayan, Ning, and Mehr. It is a cubic B-Spline parameterization based on leading edge radius, trailing edge camber and wedge angle, and optional trailing edge gap distance.

FLOWFoil.AirfoilTools.BasicBSpline — Type

Fields:

leading_edge_radius::Float : leading edge radius

trailing_edge_camber_angle::Float : trailing edge camber angle (angle of chordline from horizontal at trailing edge).

wedge_angle::Float : Wedge angle (angle between upper and lower surfaces at trailing edge).

trailing_edge_gap::Float=0.0 : distance between upper and lower surfaces at trailing edge. A value of zero indicates a sharp trailing edge.

third_ctrlpt_position::Float=1.0/3.0 : the position of the third control point. This is an inherent value in the parameterization and if changed, the other parameters will not behave as they are defined here.

A basic B-Spline airfoil can be generated with

FLOWFoil.AirfoilTools.basic_bspline — Function

basic_bspline(parameters::BasicBSpline; N=160, split=false, return_nurbs=false)

Obtain airfoil coordinates from a B-Spline parameterization method.

Arguments:

parameters::BasicBSpline : BasicBSpline parameters.

Keyword Arguments:

N::Integer=160 : number of points to use when defining the airfoil

split::Bool : flag whether to output upper and lower coordinates separately

return_nurbs::Bool : flag whether to output spline knots and control points as well

Returns:

if split=false

x::AbstractArray{Float} : x-coordinates from lower TE clockwise to upper TE

z::AbstractArray{Float} : z-coordinates from lower TE clockwise to upper TE

if split=true

xu::AbstractArray{Float} : array of x-coordinates for the upper half of the airfoil (LE to TE)

zu::AbstractArray{Float} : array of z-coordinates for the upper half of the airfoil (LE to TE)

xl::AbstractArray{Float} : array of x-coordinates for the lower half of the airfoil (LE to TE)

zl::AbstractArray{Float} : array of z-coordinates for the lower half of the airfoil (LE to TE)

if return_nurbs=true, also return:

NURBSu::NURBS.NURBScurve : upper spline object

NURBSl::NURBS.NURBScurve : lower spline object

gbs(leading_edge_radius, trailing_edge_camber_angle, wedge_angle; perturbations=nothing, trailing_edge_gap=0, degree=3, third_ctrlpt_position=1/3, weights=nothing, split=false, return_nurbs=false)

Obtain airfoil coordinates from a B-Spline parameterization method.

Arguments:

leading_edge_radius::Float : Leading Edge Radius

trailing_edge_camber_angle::Float : Trailing Edge Camber Angle (degrees)

wedge_angle::Float : Wedge Angle (degrees)

Keyword Arguments:

trailing_edge_gap::Float : Trailing Edge Gap

third_ctrlpt_position::Float : The x postion of the third control point.

split::Bool : flag whether to output upper and lower coordinates separately

return_nurbs::Bool : flag whether to output spline object

Returns:

if split=false

x::AbstractArray{Float} : x-coordinates from lower TE clockwise to upper TE

z::AbstractArray{Float} : z-coordinates from lower TE clockwise to upper TE

if split=true

xu::AbstractArray{Float} : array of x-coordinates for the upper half of the airfoil (LE to TE)

zu::AbstractArray{Float} : array of z-coordinates for the upper half of the airfoil (LE to TE)

xl::AbstractArray{Float} : array of x-coordinates for the lower half of the airfoil (LE to TE)

zl::AbstractArray{Float} : array of z-coordinates for the lower half of the airfoil (LE to TE)

if return_nurbs=true, also return:

NURBSu::NURBS.NURBScurve : upper spline object

NURBSl::NURBS.NURBScurve : lower spline object

Parametric Section (PARSEC)

Standard

The nominal PARSEC parameter type implemented is defined as

FLOWFoil.AirfoilTools.PARSEC — Type

Fields:

leading_edge_radius::Float : leading edge radius

maximum_thickness_xu::Float : x-position of maximum thickness for upper side.

maximum_thickness_xl::Float : x-position of maximum thickness for lower side.

maximum_thickness_zu::Float : value of maximum thickness (from zero) for upper side.

maximum_thickness_zl::Float : value of maximum thickness (from zero) for lower side.

curvature_u::Float : curvature at point of maximum thickness on upper side.

curvature_l::Float : curvature at point of maximum thickness on lower side.

trailing_edge_angle::Float : angle from chordline to horizontal at trailing edge.

boattail_angle::Float : angle from chordline to upper/lower surfaces (half of wedge angle).

trailing_edge_gap::Float=0.0 : total gap distance between upper and lower surfaces at treailing edge.

trailing_edge_z::Float=0.0 : z-position of midpoint between upper and lower surfaces at trailing edge.

A PARSEC airfoil can be generated with

FLOWFoil.AirfoilTools.parsec — Function

parsec(p::PARSEC; N::Integer=80, split=false)

Calculate the x,z airfoil coordinates for both top and bottom surfaces using standard PARSEC Parameterization method.

Use parsec() for modified PARSEC implementation.

Arguments:

p::PARSEC : PARSEC paramters including:

leading_edge_radius : Leading edge radius

maximum_thickness_xu : chordwise position of maximum thickness of upper side

maximum_thickness_xl : chordwise position of maximum thickness of lower side

maximum_thickness_zu : z-coordinate at maximum thickness of upper side

maximum_thickness_zl : z-coordinate at maximum thickness of lower side

curvature_u : second derivative of surface geometry at maximum thickness of upper side

curvature_l : second derivative of surface geometry at maximum thickness of lower side

trailing_edge_angle : trailing edge angle

boattail_angle : boat-tail angle

trailing_edge_gap : z-position of center of trailing edge

trailing_edge_z : z-distance between upper and lower surface trailing edge points

Keyword Arguments:

N::Integer=80 : Number of x stations along chord

split::Bool : Flag wheter to split into upper and lower halves.

Returns:

If split == false:

x::AbstractArray{Float} : Vector of x coordinates, clockwise from trailing edge.

z::AbstractArray{Float} : Vector of z coordinates, clockwise from trailing edge.

If split == true:

xl::AbstractArray{Float} : Vector of lower half of x coordinates from trailing edge to leading edge.

xu::AbstractArray{Float} : Vector of upper half of x coordinates from leading edge to trailing edge.

zl::AbstractArray{Float} : Vector of lower half of z coordinates from trailing edge to leading edge.

zu::AbstractArray{Float} : Vector of upper half of z coordinates from leading edge to trailing edge.

parsec(p::AbstractArray{Float}; N::Integer=80, split=false)

Calculate the x,z airfoil coordinates for both top and bottom surfaces using standard PARSEC Parameterization method.

Use parsec() for modified PARSEC implementation.

Arguments:

p::AbstractArray{Float} : PARSEC paramters including:

leading_edge_radius : Leading edge radius

maximum_thickness_xu : chordwise position of maximum thickness of upper side

maximum_thickness_xl : chordwise position of maximum thickness of lower side

maximum_thickness_zu : z-coordinate at maximum thickness of upper side

maximum_thickness_zl : z-coordinate at maximum thickness of lower side

curvature_u : second derivative of surface geometry at maximum thickness of upper side

curvature_l : second derivative of surface geometry at maximum thickness of lower side

trailing_edge_angle : trailing edge angle

boattail_angle : boat-tail angle

trailing_edge_gap : z-position of center of trailing edge

trailing_edge_z : z-distance between upper and lower surface trailing edge points

Keyword Arguments:

N::Integer=80 : Number of x stations along chord

split::Bool : Flag wheter to split into upper and lower halves.

Returns:

If split == false:

x::AbstractArray{Float} : Vector of x coordinates, clockwise from trailing edge.

z::AbstractArray{Float} : Vector of z coordinates, clockwise from trailing edge.

If split == true:

xl::AbstractArray{Float} : Vector of lower half of x coordinates from trailing edge to leading edge.

xu::AbstractArray{Float} : Vector of upper half of x coordinates from leading edge to trailing edge.

zl::AbstractArray{Float} : Vector of lower half of z coordinates from trailing edge to leading edge.

zu::AbstractArray{Float} : Vector of upper half of z coordinates from leading edge to trailing edge.

For a given set of x,z coordinates a best-fit PARSEC airfoil can be found using

FLOWFoil.AirfoilTools.determine_parsec — Function

determine_parsec(x,z)

Uses LsqFit to go from x-z coordinates to standard PARSEC parameters.

Returns:

parameters::PARSEC : an parameter object of type PARSEC

Modified

Also implemented in AirfoilTools is a modified PARSEC parameterization that give direct control to the trailing edge surfaces of the upper and lower sides. The ModifiedPARSEC type is defined as

FLOWFoil.AirfoilTools.ModifiedPARSEC — Type

Fields:

leading_edge_radius::Float : leading edge radius

maximum_thickness_xu::Float : x-position of maximum thickness for upper side.

maximum_thickness_xl::Float : x-position of maximum thickness for lower side.

maximum_thickness_zu::Float : value of maximum thickness (from zero) for upper side.

maximum_thickness_zl::Float : value of maximum thickness (from zero) for lower side.

curvature_u::Float : curvature at point of maximum thickness on upper side.

curvature_l::Float : curvature at point of maximum thickness on lower side.

trailing_edge_tangent_u::Float : angle of surface at upper side trailing edge.

trailing_edge_tangent_l::Float : angle of surface at lower side trailing edge.

trailing_edge_zu::Float=0.0 : z-position of upper side trailing edge.

trailing_edge_zl::Float=0.0 : z-position of lower side trailing edge.

A Modified PARSEC airfoil can be generated with

FLOWFoil.AirfoilTools.modified_parsec — Function

parsec(p::ModifiedPARSEC; N::Int=80, split=false)

Calculate the x,y airfoil coordinates for both top and bottom surfaces using modified PARSEC Parameterization method.

Use parsec() for standard PARSEC implementation. This modified version employs direct values for trailing edge position and angles for each surface.

Arguments:

p::ModifiedPARSEC : ModifiedPARSEC paramters including:

leading_edge_radius : Leading edge radius

maximum_thickness_xu : chordwise position of maximum thickness of upper side

maximum_thickness_xl : chordwise position of maximum thickness of lower side

maximum_thickness_zu : z-coordinate at maximum thickness of upper side

maximum_thickness_zl : z-coordinate at maximum thickness of lower side

curvature_u : second derivative of surface geometry at maximum thickness of upper side

curvature_l : second derivative of surface geometry at maximum thickness of lower side

trailing_edge_tangent_u : trailing edge tangent angle of upper side

trailing_edge_tangent_l : trailing edge tangent angle of lower side

trailing_edge_zu : z-position of trailing edge of upper side

trailing_edge_zl : z-position of trailing edge of lower side

Keyword Arguments:

N::Integer=80 : Number of x stations along chord

split::Bool : Flag wheter to split into upper and lower halves.

Returns:

If split == false:

x::AbstractArray{Float} : Vector of x coordinates, clockwise from trailing edge.

z::AbstractArray{Float} : Vector of z coordinates, clockwise from trailing edge.

If split == true:

xl::AbstractArray{Float} : Vector of lower half of x coordinates from trailing edge to leading edge.

xu::AbstractArray{Float} : Vector of upper half of x coordinates from leading edge to trailing edge.

zl::AbstractArray{Float} : Vector of lower half of z coordinates from trailing edge to leading edge.

zu::AbstractArray{Float} : Vector of upper half of z coordinates from leading edge to trailing edge.

parsec(p::AbstractArray{Float}; N::Int=80, split=false)

Calculate the x,y airfoil coordinates for both top and bottom surfaces using modified PARSEC Parameterization method.

Use parsec() for standard PARSEC implementation. This modified version employs direct values for trailing edge position and angles for each surface.

Arguments:

p::AbstractArray{Float} : ModifiedPARSEC paramters including:

leading_edge_radius : Leading edge radius

maximum_thickness_xu : chordwise position of maximum thickness of upper side

maximum_thickness_xl : chordwise position of maximum thickness of lower side

maximum_thickness_zu : z-coordinate at maximum thickness of upper side

maximum_thickness_zl : z-coordinate at maximum thickness of lower side

curvature_u : second derivative of surface geometry at maximum thickness of upper side

curvature_l : second derivative of surface geometry at maximum thickness of lower side

trailing_edge_tangent_u : trailing edge tangent angle of upper side

trailing_edge_tangent_l : trailing edge tangent angle of lower side

trailing_edge_zu : z-position of trailing edge of upper side

trailing_edge_zl : z-position of trailing edge of lower side

Keyword Arguments:

N::Integer=80 : Number of x stations along chord

split::Bool : Flag wheter to split into upper and lower halves.

Returns:

If split == false:

x::AbstractArray{Float} : Vector of x coordinates, clockwise from trailing edge.

z::AbstractArray{Float} : Vector of z coordinates, clockwise from trailing edge.

If split == true:

xl::AbstractArray{Float} : Vector of lower half of x coordinates from trailing edge to leading edge.

xu::AbstractArray{Float} : Vector of upper half of x coordinates from leading edge to trailing edge.

zl::AbstractArray{Float} : Vector of lower half of z coordinates from trailing edge to leading edge.

zu::AbstractArray{Float} : Vector of upper half of z coordinates from leading edge to trailing edge.

For a given set of x,z coordinates a best-fit Modified PARSEC airfoil can be found using

FLOWFoil.AirfoilTools.determine_modified_parsec — Function

determine_modified_parsec(x,z)

Uses LsqFit to go from x-z coordinates to modified ModifiedPARSEC parameters.

Returns:

parameters::ModifiedPARSEC : an parameter object of type ModifiedPARSEC

Contributing

We welcome the addition of other common parameterizations. Additions should have outputs consistent with current parameterizations.