Basic Tutorials

FLOWFoil includes various panel method implementations that are available based on the method keyword argument. Here we go over the available methods and their various options

Xfoil Method

An Xfoil-like method, actually based on mfoil can be accessed using the Mfoil method type:

FLOWFoil.Mfoil — Type

Mfoil <: Method

Fields:

viscous::Bool : Flag whether to run an inviscid (false) or viscous (true) analysis.

Note that we have also set Xfoil=Mfoil so you can also use the Xfoil method type with identical results. Currently, this method only includes the inviscid parts of Xfoil/Mfoil, so Reynolds and Mach number inputs do nothing if used.

using FLOWFoil

x, y = AirfoilTools.naca4()

angles_of_attack = range(-5.0, 15.0; step=1)

# viscous solver not yet implemented
method = Mfoil(viscous=false)

outputs = analyze(x, y, angles_of_attack; method=method)

FLOWFoil.InviscidOutputs{Matrix{Float64}, Matrix{Float64}}([-0.692908072768126 -0.6934105709397785 … -0.6665000542362963 -0.6630363805906164; -0.7421458177430701 -0.7424731035417773 … -0.7099266288461078 -0.7060179988578108; … ; 0.7437749192650895 0.7445314515350743 … 0.7194790365407877 0.7159657903190176; 0.692908072768126 0.6934105709397785 … 0.6665000542362963 0.6630363805906164], [0.5198784026927613 0.5191817801089704 … 0.5557776777030141 0.5603827580132952; 0.44921958520646976 0.44873369051704126 … 0.49600418165520066 0.5015385852888123; … ; 0.4467988694722095 0.44567291767507533 … 0.48234991597833987 0.4873929870928645; 0.5198784026927613 0.5191817801089704 … 0.5557776777030141 0.5603827580132952], [-0.34847343924627516; -0.22779847891023397; … ; 1.9206359736335232; 2.0363902468753956;;], [0.0005166916701355751; 0.00048227845938455914; … ; 0.0008839264682029976; 0.0009596330988951767;;], [-0.048888114413008436; -0.05019850205092104; … ; -0.07593956183481718; -0.07737833053995849;;])

Lewis' Method for Axisymmetric Bodies

An axisymmetric method based on that described by Lewis can be accessed using the Lewis method type:

FLOWFoil.Lewis — Type

Lewis <: Method

Fields:

body_of_revolution::AbstractVector{Bool} : Flag(s) whether bodies are bodies of revolutions or not (false indicates an annular airfoil)

Note that if multiple bodies are used, the annular airfoil should come before the body of revolution.

using FLOWFoil

x, r = AirfoilTools.naca4()

# give the duct some diameter (see note below)
r .+= 1.0

# indicate that the body is not a body of revolution (i.e. a duct)
method = Lewis(; body_of_revolution=[false])

# note: we need to input an an angle of attack, even though it is unused
outputs = analyze(x, r, [0.0]; method=method)

FLOWFoil.InviscidOutputs{Matrix{Float64}, Matrix{Float64}}([-0.7169024195697359; -0.7767614729891681; … ; 0.7525392894347234; 0.7169024195697359;;], [0.4860509208150584; 0.3966416140796979; … ; 0.4336846178570817; 0.4860509208150584;;], [0.0;;], [0.0;;], [0.0;;])

Note

No part of the geometry for an axisymmetric body can reside below z=0, otherwise an error will be thrown.

Martensen's Method for Periodic Bodies (Cascades)

A periodic method for cascade analysis based on that developed by Martensen can be accessed using the Martensen method type:

FLOWFoil.Martensen — Type

Martensen <: Method

Fields:

cascade::Bool=true : flag to apply cascade treatment or not

solidity::Float=0.0 : Ratio between airfoil chord length and pitch. Airfoil pitch is simply the distance between chordlines in the cascade.

stagger::Float=0.0 : Angle (in degrees) from axis of ration to airfoil chordline. Note that stagger is equivalent to the inflow angle minus the angle of attack.

transition_value::Float=Inf : pitchtochord ratio at which we stop applying cascade effects (Lewis uses 30 in his implementation)

curvature_correction::Bool=false : flag to apply curvature correction from Lewis

Note

If the cascade option is set to false, this method becomes a standard planar airfoil method, but uses constant rather than linear vortices, so the Mfoil/Xfoil method is superior in that case.

using FLOWFoil

x, y = AirfoilTools.naca4()

angles_of_attack = range(-5.0, 15.0; step=1)

# The cascade method requires solidity (closeness) of sections and stagger (inflow angle - angle of attack)
method = Martensen(solidity=1.2, stagger=15.0)

outputs = analyze(x, y, angles_of_attack; method=method)

FLOWFoil.InviscidOutputs{Matrix{Float64}, Matrix{Float64}}([-0.7544599115670336 -0.7538202145476088 … -0.7035661149684271 -0.6986759587790523; -0.8179805047360362 -0.8172754670028922 … -0.7625875831070599 -0.757274883525712; … ; 0.7914518400493568 0.7907925155060266 … 0.738281627439637 0.7331627666149593; 0.8179805047360362 0.8172754670028922 … 0.7625875831070599 0.757274883525712], [0.43079024183826375 0.431755084139397 … 0.504994721868234 0.5118519046241721; 0.33090789387177943 0.33206081103520446 … 0.41846017809093305 0.42653475078111924; … ; 0.37360398488248736 0.3746471974196507 … 0.45494023858508104 0.4624723576494988; 0.33090789387177943 0.33206081103520446 … 0.41846017809093305 0.42653475078111924], [-0.09365805378538773; -0.06445079787741738; … ; 0.4651696039312523; 0.49552656511152676;;], [0.0; 0.0; … ; 0.0; 0.0;;], [0.0; 0.0; … ; 0.0; 0.0;;])

Hess-Smith 2D Method for Educational Purposes

We also have a version of the Hess-Smith method primarily for educational use that can be accessed with the HessSmith method type:

FLOWFoil.HessSmith — Type

HessSmith <: Method

Fields:

V_inf::TF=1.0 : magnitude of the free stream velocity

using FLOWFoil

x, y = AirfoilTools.naca4()

angles_of_attack = range(-5.0, 15.0; step=1)

method = HessSmith(V_inf=1.0)

outputs = analyze(x, y, angles_of_attack; method=method)

FLOWFoil.InviscidOutputs{Matrix{Float64}, Matrix{Float64}}([-0.7946235625220246 -0.7690972535232394 … -0.2774052102761565 -0.2488267826293874; -0.8105007640203398 -0.7983289973602731 … -0.5415865323798975 -0.5255069094877589; … ; 0.7394470657805323 0.7527187232719785 … 0.9488396831485161 0.9571487549581185; 0.6411708712808938 0.6686331523271828 … 1.1203593016418056 1.1426039960901302], [0.36857339388480603 0.4084894146230099 … 0.9230463493116414 0.9380852322463076; 0.34308851152244546 0.36267081197374107 … 0.7066840279447182 0.7238424880806243; … ; 0.4532180369085611 0.4334145236358026 … 0.09970325568262361 0.08386626088212368; 0.5888999138208995 0.5529297076090144 … -0.2552049647753145 -0.3055438918811342], [-0.36371620484454154; -0.2391879772820596; … ; 1.9797263356391366; 2.0993267598115133;;], [0.0; 0.0; … ; 0.0; 0.0;;], [0.0; 0.0; … ; 0.0; 0.0;;])