Basic Tutorials

FLOWFoil includes various panel method implementations that are available based on the method keyword argument. Here we go over the available methods and their various options

Xfoil Method

An Xfoil-like method, actually based on mfoil can be accessed using the Mfoil method type:

FLOWFoil.Mfoil — Type

Mfoil <: Method

Fields:

viscous::Bool : Flag whether to run an inviscid (false) or viscous (true) analysis.

Note that we have also set Xfoil=Mfoil so you can also use the Xfoil method type with identical results. Currently, this method only includes the inviscid parts of Xfoil/Mfoil, so Reynolds and Mach number inputs do nothing if used.

using FLOWFoil

x, y = AirfoilTools.naca4()

angles_of_attack = range(-5.0, 15.0; step=1)

# viscous solver not yet implemented
method = Mfoil(viscous=false)

outputs = analyze(x, y, angles_of_attack; method=method)

InviscidOutputs{Matrix{Float64}, Matrix{Float64}}([-0.692908072768126 -0.6934105709397785 … -0.6665000542362963 -0.6630363805906164; -0.7421458177430701 -0.7424731035417773 … -0.7099266288461078 -0.7060179988578108; … ; 0.7437749192650895 0.7445314515350743 … 0.7194790365407877 0.7159657903190176; 0.692908072768126 0.6934105709397785 … 0.6665000542362963 0.6630363805906164], [0.5198784026927613 0.5191817801089704 … 0.5557776777030141 0.5603827580132952; 0.44921958520646976 0.44873369051704126 … 0.49600418165520066 0.5015385852888123; … ; 0.4467988694722095 0.44567291767507533 … 0.48234991597833987 0.4873929870928645; 0.5198784026927613 0.5191817801089704 … 0.5557776777030141 0.5603827580132952], [-0.34847343924627516; -0.22779847891023397; … ; 1.9206359736335232; 2.0363902468753956;;], [0.0005166916701355751; 0.00048227845938455914; … ; 0.0008839264682029976; 0.0009596330988951767;;], [-0.048888114413008436; -0.05019850205092104; … ; -0.07593956183481718; -0.07737833053995849;;])

Lewis' Method for Axisymmetric Bodies

An axisymmetric method based on that described by Lewis can be accessed using the Lewis method type:

FLOWFoil.Lewis — Type

Lewis <: Method

Fields:

body_of_revolution::AbstractVector{Bool} : Flag(s) whether bodies are bodies of revolutions or not (false indicates an annular airfoil)

Note that if multiple bodies are used, the annular airfoil should come before the body of revolution.

using FLOWFoil

x, r = AirfoilTools.naca4()

# give the duct some diameter (see note below)
r .+= 1.0

# indicate that the body is not a body of revolution (i.e. a duct)
method = Lewis(; body_of_revolution=[false])

# note: we need to input an an angle of attack, even though it is unused
outputs = analyze(x, r, [0.0]; method=method)

InviscidOutputs{Matrix{Float64}, Matrix{Float64}}([-0.7169024195697359; -0.7767614729891681; … ; 0.7525392894347234; 0.7169024195697359;;], [0.4860509208150584; 0.3966416140796979; … ; 0.4336846178570817; 0.4860509208150584;;], [0.0;;], [0.0;;], [0.0;;])

Note

No part of the geometry for an axisymmetric body can reside below z=0, otherwise an error will be thrown.

Martensen's Method for Periodic Bodies (Cascades)

A periodic method for cascade analysis based on that developed by Martensen can be accessed using the Martensen method type:

FLOWFoil.Martensen — Type

Martensen <: Method

Fields:

cascade::Bool=true : flag to apply cascade treatment or not

solidity::Float=0.0 : Ratio between airfoil chord length and pitch. Airfoil pitch is simply the distance between chordlines in the cascade.

stagger::Float=0.0 : Angle (in degrees) from axis of ration to airfoil chordline. Note that stagger is equivalent to the inflow angle minus the angle of attack.

transition_value::Float=Inf : pitchtochord ratio at which we stop applying cascade effects (Lewis uses 30 in his implementation)

curvature_correction::Bool=false : flag to apply curvature correction from Lewis

Note

If the cascade option is set to false, this method becomes a standard planar airfoil method, but uses constant rather than linear vortices, so the Mfoil/Xfoil method is superior in that case.

using FLOWFoil

x, y = AirfoilTools.naca4()

angles_of_attack = range(-5.0, 15.0; step=1)

# The cascade method requires solidity (closeness) of sections and stagger (inflow angle - angle of attack)
method = Martensen(solidity=1.2, stagger=15.0)

outputs = analyze(x, y, angles_of_attack; method=method)

InviscidOutputs{Matrix{Float64}, Matrix{Float64}}([-0.7544599115670336 -0.7538202145476088 … -0.7035661149684271 -0.6986759587790523; -0.8179805047360362 -0.8172754670028922 … -0.7625875831070599 -0.757274883525712; … ; 0.7914518400493568 0.7907925155060266 … 0.738281627439637 0.7331627666149593; 0.8179805047360362 0.8172754670028922 … 0.7625875831070599 0.757274883525712], [0.43079024183826375 0.431755084139397 … 0.504994721868234 0.5118519046241721; 0.33090789387177943 0.33206081103520446 … 0.41846017809093305 0.42653475078111924; … ; 0.37360398488248736 0.3746471974196507 … 0.45494023858508104 0.4624723576494988; 0.33090789387177943 0.33206081103520446 … 0.41846017809093305 0.42653475078111924], [-0.09365805378538773; -0.06445079787741738; … ; 0.4651696039312523; 0.49552656511152676;;], [0.0; 0.0; … ; 0.0; 0.0;;], [0.0; 0.0; … ; 0.0; 0.0;;])

Hess-Smith 2D Method

We also have a version of the Hess-Smith method primarily for educational use that can be accessed with the HessSmith method type:

FLOWFoil.HessSmith — Type

HessSmith <: Method

Fields:

V_inf::TF=1.0 : magnitude of the free stream velocity

using FLOWFoil

x, y = AirfoilTools.naca4()

angles_of_attack = range(-5.0, 15.0; step=1)

method = HessSmith(V_inf=1.0)

outputs = analyze(x, y, angles_of_attack; method=method)

InviscidOutputs{Matrix{Float64}, Matrix{Float64}}([-0.7946235625220246 -0.7690972535232394 … -0.2774052102761565 -0.2488267826293874; -0.8105007640203398 -0.7983289973602731 … -0.5415865323798975 -0.5255069094877589; … ; 0.7394470657805323 0.7527187232719785 … 0.9488396831485161 0.9571487549581185; 0.6411708712808938 0.6686331523271828 … 1.1203593016418056 1.1426039960901302], [0.36857339388480603 0.4084894146230099 … 0.9230463493116414 0.9380852322463076; 0.34308851152244546 0.36267081197374107 … 0.7066840279447182 0.7238424880806243; … ; 0.4532180369085611 0.4334145236358026 … 0.09970325568262361 0.08386626088212368; 0.5888999138208995 0.5529297076090144 … -0.2552049647753145 -0.3055438918811342], [-0.36371620484454154; -0.2391879772820596; … ; 1.9797263356391366; 2.0993267598115133;;], [0.0; 0.0; … ; 0.0; 0.0;;], [0.0; 0.0; … ; 0.0; 0.0;;])

NeuralFoil Translation from Python to Julia

NeuralFoil is a multi-layer perceptron model of Xfoil. We provide a very basic translation of the get_aero_from_coordinates function from NeuralFoil via the NeuralFoil method type:

FLOWFoil.NeuralFoil — Type

NeuralFoil <: Method

In general usage, this struct is generated by calling the NeuralFoil constructor.

Fields:

mean_inputs_scaled::Vector

cov_inputs_scaled::Matrix

inv_cov_inputs_scaled::Matrix

weights::Vector{Vector}}

biases::Vector{Vector}}

Re::Float

Ma::Float

`n_crit::Float'

`xtr_upper::Float'

`xtr_lower::Float'

using FLOWFoil

x, y = AirfoilTools.naca4()

angles_of_attack = range(-5.0, 15.0; step=1)

reynolds = 2e6
mach = 0.0

method = NeuralFoil(reynolds, mach; model_size="xlarge")

outputs = analyze([x y], angles_of_attack; method=method)

NeuralOutputs{Matrix{Float64}, Vector{Float64}}([-0.2933311019283913, -0.18876019767630003, -0.08379098600652851, 0.0219572437926992, 0.1266285349182396, 0.23094070641438869, 0.33140184809833984, 0.43075801549679804, 0.5437956319188877, 0.6751252484567893  …  0.9173976116144951, 1.01127036586331, 1.0950413623630861, 1.1746293231720153, 1.2521587648796082, 1.3266013191834263, 1.394500885353656, 1.4541725821366875, 1.5065671810505203, 1.5501162499121675], [0.007068834786627557, 0.006662257257342131, 0.0063070869005499115, 0.006020143008742286, 0.005759704716589132, 0.005528653704595198, 0.005202412246418282, 0.005079913078107053, 0.005379608198706969, 0.006040227265139352  …  0.008350181908413683, 0.00953930402455891, 0.010565955134481131, 0.011651665103825525, 0.012898448909796177, 0.014304170474114562, 0.01592484879058611, 0.017917221578584306, 0.020543058013902746, 0.024203354972131996], [-0.058779886648120894, -0.05715910207254108, -0.0554365147859634, -0.05389397837115911, -0.05224289009763079, -0.05045568171125057, -0.04778032335510067, -0.044512135822076734, -0.044337419489327276, -0.048434351587123683  …  -0.05373583254234219, -0.0506366261572509, -0.045442258666632464, -0.039560084277762295, -0.03344478166897038, -0.026999064962783825, -0.01987615008342316, -0.012267191079644213, -0.005148913124969967, 0.0008555939939929536], [0.9658690711181539, 0.9652067418272086, 0.9617760613615883, 0.9584140767926261, 0.9567722239783185, 0.9590533634437008, 0.9636621596081686, 0.9608110453659165, 0.9531057367694116, 0.9531960172267077  …  0.9545028623785625, 0.9604992091254606, 0.9660191804500236, 0.968533079630373, 0.9694143916152298, 0.9701851070191954, 0.9712104746033379, 0.9723750731359512, 0.973319631587101, 0.9736220088414475], [0.857936898969106, 0.8082819228401218, 0.7445617540900145, 0.685871695687948, 0.6309780193166257, 0.5710280976118899, 0.5126745137524353, 0.4597665753058441, 0.4030528616118132, 0.3351543081228706  …  0.14434130816742888, 0.08381936421372918, 0.05682724249545559, 0.041834702961313364, 0.031219459768666413, 0.023719879840444213, 0.018248534268198724, 0.013966615632273471, 0.01094266299281696, 0.00993066528554784], [0.06397680009584872, 0.08783556137066384, 0.13654757154012803, 0.2096178770675631, 0.3191333199112868, 0.46762021411932087, 0.68460999196192, 0.834669094465348, 0.9698391718017733, 1.0  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.3819029397297302 0.5024393403502363 … 2.5670318987169964 2.642724798101353; 0.753532234207844 0.8358492733540319 … 2.16053007619928 2.2094678576933573; … ; 0.9581681520036708 0.9604121021696361 … 0.9628583948019097 0.9915951664733093; 0.9134422814515906 0.9116082621483093 … 0.9677826037890651 0.9930711212623621], [-1.715797267894175 -1.5978982268296895 … 0.6028345148775189 0.6920491466284069; -1.5781903438258529 -1.4870287577585035 … -0.0970609265411555 -0.04019996361411715; … ; -0.9286852735340777 -0.9284385755084819 … -0.9672987988167001 -0.9787305694739549; -0.8919441638352135 -0.8909607395865817 … -0.9791610148633466 -0.9972766263179289], [4.236679152089569e-5 4.474234103419703e-5 … 7.851919294619531e-5 8.746023181065599e-5; 7.1615563672133e-5 7.31864197423974e-5 … 0.0002590369590802542 0.00027213383864823027; … ; 0.0007526744360259522 0.0008587017708668452 … 0.009577096628142748 0.009456113594909351; 0.0009544757436572158 0.001085624604756605 … 0.009513641293196684 0.009539056153836444], [2.2538678435802355 2.2700200452912442 … 2.799988125869715 2.5337782396543274; 2.3362057787559567 2.355185955203134 … 1.3791014982387557 1.3975116678893968; … ; 1.4745407936112984 1.3748311857018445 … 4.188556625410758 5.3926265027672216; 1.4388064551193394 1.472760427908293 … 4.794228645493481 5.970227552393945], [5.690123624861929e-5 5.706270969830473e-5 … 4.223923938089828e-5 4.034541289854516e-5; 9.445621645942297e-5 9.425896058053627e-5 … 7.740588903942295e-5 0.0001624515869566382; … ; 0.003629700215084911 0.0032136441848326786 … 0.00031588522980340055 0.0002993235747445111; 0.004224635661083872 0.00375467430849771 … 0.0003197732195297446 0.0002969637737584136], [2.7431906296454716 2.5912712029024036 … 2.191167056015233 2.1877744833060224; 3.4385826471224643 3.2794182190818173 … 2.250665851407489 2.249160139994979; … ; 1.555549105457811 1.5262948826309695 … 2.065288890046051 2.060986963948944; 1.6237392987176371 1.5950535696365467 … 2.0962920931615714 2.061729050290894])

Note that the NeuralFoil method does not allow multi-body analysis like the other methods do as it is based specifically on Xfoil. We also return a separate output type for the NeuralFoil method:

FLOWFoil.NeuralOutputs — Type

NeuralOutputs

Fields

cl::Vector : lift coefficients

cd::Vector : drag coefficients

cm::Vector : moment coefficients

confidence::Vector : confidence factor reported by NeuralFoil

top_xtr::Vector : laminar to turbulent transition location of top surface

bot_xtr::Vector : laminar to turbulent transition location of bottom surface

upper_bl_ue_over_vinf::Matrix : upper boundary layer normalized surface velocity

lower_bl_ue_over_vinf::Matrix : lower boundary layer normalized surface velocity

upper_theta::Matrix : upper boundary layer momentum thickness

upper_H::Matrix : upper boundary layer shape factor

lower_theta::Matrix : lower boundary layer momentum thickness

lower_H::Matrix : lower boundary layer shape factor