Basic Tutorials

FLOWFoil includes various panel method implementations that are available based on the method keyword argument. Here we go over the available methods and their various options

Xfoil Method

An Xfoil-like method, actually based on mfoil can be accessed using the Mfoil method type:

FLOWFoil.Mfoil — Type

Mfoil <: Method

Fields:

viscous::Bool : Flag whether to run an inviscid (false) or viscous (true) analysis.

Note that we have also set Xfoil=Mfoil so you can also use the Xfoil method type with identical results. Currently, this method only includes the inviscid parts of Xfoil/Mfoil.

using FLOWFoil

x, y = AirfoilTools.naca4()

angles_of_attack = range(-5.0, 15.0, step=1)

method = Mfoil()

outputs = AirfoilTools.analyze(x, y, angles_of_attack; method=method)

Lewis' Method for Axisymmetric Bodies

A axisymmetric method based on that described by Lewis can be accessed using the Lewis method type:

FLOWFoil.Lewis — Type

Lewis <: Method

Fields:

body_of_revolution::AbstractVector{Bool} : Flag(s) whether bodies are bodies of revolutions or not (false indicates an annular airfoil)

Note that if multiple bodies are used, the annular airfoil should come before the body of revolution.

using FLOWFoil

# - DUCT - #

x, r = AirfoilTools.naca4()

# give the duct some diameter
r .+= 1.0

# indicate that the body is not a body of revolution (lying on the axis)
method = Lewis(body_of_revolution = [false])

# angle of attack defaults to zero, which is what we want for the axisymmetric case
outputs = analyze(x, r; method=method)

(tangential_velocities = [[-0.7169024195697359, -0.7767614729891681, -0.814507933798744, -0.8385200094270594, -0.8550887697044678, -0.867552290803076, -0.877720452235311, -0.8865349233940396, -0.8944648542630386, -0.9017444243879593  …  0.9207112114337079, 0.9073350136053355, 0.8925021164495869, 0.8758333824106442, 0.856896268076917, 0.8352910124763073, 0.8107861589603169, 0.7833431838198711, 0.7525392894347234, 0.7169024195697359]], surface_pressures = [[0.4860509208150584, 0.3966416140796979, 0.3365768257789009, 0.2968841937904442, 0.2688231959252997, 0.24735302272233495, 0.22960680772784103, 0.2140558296027243, 0.19993262448820115, 0.18685699308522796  …  0.152290865140274, 0.17674317308580556, 0.20343997213300802, 0.23291588625513027, 0.26572878575585246, 0.3022889244763054, 0.34262580443837576, 0.3863734563629476, 0.4336846178570817, 0.4860509208150584]])

Note

No part of the geometry for an axisymmetric body can reside below z=0, otherwise an error will be thrown.

Martensen's Method for Periodic Bodies (Cascades)

A periodic method for cascade analysis based on that developed by Martensen can be accessed using the Martensen method type:

FLOWFoil.Martensen — Type

Martensen{} <: Method

Fields:

cascade::Bool=true : flag to apply cascade treatment or not

solidity::Float=0.0 : Ratio between airfoil chord length and pitch. Airfoil pitch is simply the distance between chordlines in the cascade.

stagger::Float=0.0 : Angle (in radians) from axis of ration to airfoil chordline

transition_value::Float=Inf : pitchtochord ratio at which we stop applying cascade effects (Lewis uses 30 in his implementation)

curvature_correction::Bool=false : flag to apply curvature correction from Lewis

using FLOWFoil

# - DUCT - #

x, y = AirfoilTools.naca4(6,4,12)

inflow_angles = range(-5.0, 15.0, step=1)

method = Martensen(solidity=0.5, stagger=30.0*pi/180.0)

outputs = analyze(x, y, inflow_angles; method=method)

Hess-Smith 2D Method for Educational Purposes

We also have a version of the Hess-Smith method primarily for educational use that can be accessed with the HessSmith method type:

FLOWFoil.HessSmith — Type

HessSmith <: Method

using FLOWFoil

# - DUCT - #

x, y = AirfoilTools.naca4(6,4,12)

angles_of_attack = range(-5.0, 15.0, step=1)

method = HessSmith()

outputs = analyze(x, y, angles_of_attack; method=method)