Quick Start

Running FLOWFoil can be done simply with a single method: analyze

FLOWFoil.analyze — Function

analyze(coordinates, flow_angles=0.0, reynolds=1e6, mach=0.0; method::Method=Mfoil())
analyze(x, y, flow_angles=0.0, reynolds=1e6, mach=0.0; method::Method=Mfoil())

Convenience function for setting up, solving, and post-processing airfoils and airfoil systems.

Arguments

coordinates::NTuple{Matrix{Float}} : Tuple of [x y] matrices of airfoil coordinates (may be a single matrix as well)

flow_angles::Vector{Float} : Vector of angles of attack (may be a single float as well)

OR

x::Vector{Float} : Vector of x-coordinates of airfoil geometry

y::Vector{Float} : Vector of y-coordinates of airfoil geometry

flow_angles::Vector{Float} : Vector of angles of attack (may be a single float as well)

Note that inputting separate vectors for airfoil coordinates is only available for analysis of single airfoils/bodies. Multi-airfoil/body systems require the use of a tuple of matrices for coordinate inputs.

Keyword Arguments

method::MethodType : desired method for solving

Returns

outputs::OutputType : outputs object (note that only inviscid methods are currently implemented)

As an introductory example, we will do a quick analysis of a NACA 2412 airfoil, with coordinates from one of the available methods in the Airfoil Tools sub-module.

Note

For any airfoil coordinate generation method (from FLOWFoil or otherwise), the coordinates must start at the trailing edge, and proceed clockwise around the airfoil. For axisymmetric bodies of revolution, coordiantes should also proceed clockwise, starting at the leading edge.

using FLOWFoil

# note: 2412 is default for the NACA 4-series implemenation
x, y = AirfoilTools.naca4()

# choose one or more angles of attack
angles_of_attack = range(-5.0, 15.0, step=1)

# analyze
outputs = analyze(x, y, angles_of_attack)

InviscidOutputs{Matrix{Float64}, Matrix{Float64}}([-0.692908072768126 -0.6934105709397785 … -0.6665000542362963 -0.6630363805906164; -0.7421458177430701 -0.7424731035417773 … -0.7099266288461078 -0.7060179988578108; … ; 0.7437749192650895 0.7445314515350743 … 0.7194790365407877 0.7159657903190176; 0.692908072768126 0.6934105709397785 … 0.6665000542362963 0.6630363805906164], [0.5198784026927613 0.5191817801089704 … 0.5557776777030141 0.5603827580132952; 0.44921958520646976 0.44873369051704126 … 0.49600418165520066 0.5015385852888123; … ; 0.4467988694722095 0.44567291767507533 … 0.48234991597833987 0.4873929870928645; 0.5198784026927613 0.5191817801089704 … 0.5557776777030141 0.5603827580132952], [-0.34847343924627516; -0.22779847891023397; … ; 1.9206359736335232; 2.0363902468753956;;], [0.0005166916701355751; 0.00048227845938455914; … ; 0.0008839264682029976; 0.0009596330988951767;;], [-0.048888114413008436; -0.05019850205092104; … ; -0.07593956183481718; -0.07737833053995849;;])