API Reference

Additional Types and Functions not already shown in other parts of the documentation.

Public

Geometry and Meshing

FLOWFoil.BodyMesh — Type

BodyMesh{TF,TB}

Mesh for single body.

Fields:

nodes::Array{Array{Float,2}} : [x y] node (panel edge) locations for airfoil

chord::Float : airfoil chord length

blunt_te::Bool : boolean for whether or not the trailing edge is blunt or not.

trailing_edge_gap::Float : trailing edge gap distance

tdp::Float : dot product of unit vectors of trailing edge bisection and gap vectors

txp::Float : pseudo-cross product of unit vectors of trailing edge bisection and gap vectors

Assuptions:

x and y coordinates start at the bottom trailing edge and proceed clockwise.

FLOWFoil.BodyMeshSystem — Type

BodyMeshSystem{TF}

System of meshes to solve.

Fields:

meshes::Array{Mesh} : Array of mesh objects.

scales::Vector{Float} : Airfoil scaling factors.

angles::Vector{Float} : Airfoil angles of attack.

locations::Array{Array{TF}} : Array of leading edge locations.

FLOWFoil.position_meshes! — Function

position_meshes!(meshsystem)

Identical to position_meshes!, but taking the inputs in as a BodyMeshSystem object.

Arguments:

meshsystem::BodyMeshSystem : Mesh system object to position.

Problems and Solutions

FLOWFoil.Problem — Type

Problem{TF,TB}

Problem definition (geometry, operating point(s), and method selection) and output behavior.

Fields:

meshes::Array{BodyMesh} : Array of mesh objects

angleofattack::Float : angle of attack to analyze.

reynolds::Float : Reynolds number to analyze.

mach::Float : Mach number to analyze.

viscous::Bool : Flag to solve viscous or inviscid only

verbose::Bool : Flag to print out verbose statements

debug::Bool : Flag to save the system structs, etc.

FLOWFoil.InviscidSolution — Type

InviscidSolution{TM,TF,TD}

Fields:

mesh::BodyMesh : BodyMesh object describing airfoil nodes etc.

panelgammas::Array{Float,2} : $\gamma_0$ and $\gamma_{90}$ values at each airfoil node.

psi0::Array{Float} : $\Psi_0$ (constant stream function) 0 and 90 values.

Ns::Array{Float} : Array of numbers of nodes for each airfoil in the system.

debug::Debug : Debug object (or nothing) depending on debug flag in Problem object.

Post Processing

FLOWFoil.Polar — Type

Polar{TF}

Fields:

lift::Float : Lift Coefficient.

drag::Float : Total Drag Coefficient.

pdrag::Float : Pressure Drag Coefficient.

idrag::Float : Induced Drag Coefficient.

moment::Float : Moment Coefficient.

surfacevelocity::Vector{Float} : surface velocity distribution

surfacepressure::Vector{Float} : surface pressure distribution

FLOWFoil.calculate_stream_grid — Function

function calculate_stream_grid(problem, solution, xrange, zrange; Nx=100, Nz=100)

Calculate stream function values across x- and z-ranges.

Arguments:

problem::Problem : Problem object associated with solution object.

solution::Solution : Solution object associated with problem.

xrange::Array{Float} : minimum and maximum values of grid in x-direction, [xmin xmax]

zrange::Array{Float} : minimum and maximum values of grid in z-direction, [zmin zmax]

Keyword Arguments:

Nx::Int : Number of gridpoints in x-direction

Nz::Int : Number of gridpoints in z-direction

Returns:

xgrid::Array{Float} : Array of x-locations of gridpoints

zgrid::Array{Float} : Array of z-locations of gridpoints

streamgrid::Array{Float,2} : Matrix of stream function values at gridpoints

Convenience Functions

Conformal Mapped Airfoils

FLOWFoil.joukowsky — Function

joukowsky(center, radius; N, fortest=false, normalize=true, split=false)

Joukowsky airfoil parameterization.

Arguments:

center::Array{Float} : [x z] location of center of circle relative to origin

radius::Float : radius of circle

Keyword Arguments:

N::Int : Number of coordinates to use

fortest::Bool : Flag to output non-coordinate paramters used in 'joukowskyflow()'

normalize::Bool : Flag whether to normalize to unit chord and translate the leading edge to zero.

split::Bool : Flag wheter to split output into upper and lower surfaces.

FLOWFoil.joukowskyflow — Function

joukowskyflow(center, radius, alpha, vinf; N=360)

Calculate the analytic surface velocities and pressures as well as lift coefficient for a joukowsky airfoil.

Arguments:

center::Array{Float} : [x z] location of circle center relative to origin

radius::Float : Radius of circle

alpha::Float : Angle of attack in degrees

vinf::Float : Freestream velocity

Keyword Arguments:

N::Int : Number of coordinates to use

Returns:

vsurf::Array{Float} : Magnitude of surface velocities at the nodes

cpsurf::Array{Float} : Surface pressures at the nodes

cl::Float : Lift coefficient

FLOWFoil.karman_trefftz — Function

karman_trefftz(beta, radius, wedge; N=360, normalize=true, split=false)

Karman-Trefftz airfoil parameterization based on angle beta, raidus, and wedge angle.

Arguments:

beta::Float : angle, in radians indicating center of circle relative to origin

radius::Float : radius of circle

wedge::Float : angle, in radians, of airfoil wedge angle

Keyword Arguments:

N::Int : number of coordinates (for entire airfoil)

normalize::Bool` : Flag whether to normalize out put to unit chord and shift to have leading edge at zero.

split::Bool` : Flag wheter to split into upper and lower halves.

Returns: IF split == False:

x::Array{Float} : Array of x coordinates

z::Array{Float} : Array of z coordinates

IF split == True:

xu::Array{Float} : Array of upper half of x coordinates

xl::Array{Float} : Array of lower half of x coordinates

zu::Array{Float} : Array of upper half of z coordinates

zl::Array{Float} : Array of lower half of z coordinates

karman_trefftz(center=[-0.1 0.1], wedge=0.0; N=360, normalize=true, split=false)

Identical to karman_trefftz(beta, radius, wedge) but using center-based version.

Arguments:

center::Array{Float} : [x z] location of circle center relative to origin

wedge::Float : angle, in radians, of airfoil wedge angle

Keyword Arguments:

N::Int : number of coordinates (for entire airfoil)

normalize::Bool : Flag whether to normalize out put to unit chord and shift to have leading edge at zero.

split::Bool : Flag wheter to split into upper and lower halves.

Returns: IF split == False:

x::Array{Float} : Array of x coordinates

z::Array{Float} : Array of z coordinates

IF split == True:

xu::Array{Float} : Array of upper half of x coordinates

xl::Array{Float} : Array of lower half of x coordinates

zu::Array{Float} : Array of upper half of z coordinates

zl::Array{Float} : Array of lower half of z coordinates

NACA 4-Series

FLOWFoil.naca4 — Function

naca4(c=2.0, p=4.0, t=12.0; N=161, x=nothing, blunt_TE=false, split=false)

Compute x, y airfoil coordinates for N nodes, based on NACA 4-Series Parameterization.

Arguments:

c::Float : Maximum camber value (percent of chord)

p::Float : Position along chord (in 10ths of chord) where maximum camber lies

t::Float : Maximum thickness of airfoil in percent chord

Keyword Arguments:

N::Int : Total number of coordinates to use (should be odd)

x::Array{Float} : x coordinates (cosine spaced coordinates used by default)

blunt_TE::Bool : Flag whether trailing edge is blunt or not

split::Bool : Flag wheter to split into upper and lower halves.

Private

Inviscid System Construction

Setup

FLOWFoil.size_system — Function

size_system(meshsystem)

Count size of inviscid system matrix.

Arguments:

meshsystem::Array{BodyMesh} : The system for which to calculate the linear system size.

FLOWFoil.get_offset — Function

get_offset(Ns)

Get the offset values for the mesh system to be used in the system matrix assembly.

Arguments:

Ns::Array{Float} : Array of numbers of nodes for each airfoil in the system.

FLOWFoil.get_trailing_edge_info — Function

get_trailing_edge_info(nodes)

Calculate various items needed for trailing edge treatment.

Arguments:

nodes::Array{Float,2} : Array of [x y] locations for the airfoil nodes.

Returns:

tdp::Float : dot product of TE bisection and TE gap unit vectors

txp::Float : "cross product" of TE bisection and TE gap unit vectors

trailing_edge_gap::Float : TE gap distance

Panel Geometry

FLOWFoil.get_r — Function

function get_r(node,point)

Calculate the vector, $\mathbf{r}$, and distance, $|r|$, from the node to the evaluation point

Arguments:

node::Array{Float} : [x y] position of node

point::Array{Float} : [x y] position of point.

Returns

r::Vector{Float} : vector from node to evaluation point

rmag::Float : length of panel between node and evaluation point

FLOWFoil.get_d — Function

get_d(node1, node2)

Calculate panel length (between adjacent nodes).

Arguments:

node1::Array{Float}(2) : [x y] location of first node

node2::Array{Float}(2) : [x y] location of second node

Returns

d::Vector{Float} : vector from node1 to node2

dmag::Float : length of panel between node1 and node2

FLOWFoil.get_theta — Function

get_theta(h, a)

Get angle (in radians) between panel and vector from node1 to evaluation point.

Arguments:

h::Float : Distance, normal to panel, between panel and evaluation point.

a::Float : Distance, tangent to panel, between node1 and evaluation point.

get_theta(h, a, dmag)

Get angle (in radians) between panel and vector from node2 to evaluation point.

Arguments:

h::Float : Distance, normal to panel, between panel and evaluation point.

a::Float : Distance, tangent to panel, between node1 and evaluation point.

dmag::Float : Panel lentgh.

FLOWFoil.get_h — Function

get_h(r1, d, dmag)

Calculate distance from panel to evalulation point in the panel normal direction.

Arguments:

r1::Vector{Float} : vector from node1 to evalulation point.

d::Vector{Float} : vector from node1 to node2.

dmag::Float : panel length

FLOWFoil.get_a — Function

get_a(r1, d, dmag)

Calculate distance from panel to evalulation point in the panel tangent direction.

Arguments:

r1::Vector{Float} : vector from node1 to evalulation point.

d::Vector{Float} : vector from node1 to node2.

dmag::Float : panel length

FLOWFoil.get_tangent — Function

get_tangent(d, dmag)

Get unit tangent to panel.

Arguments:

d::Vector{Float} : vector from node1 to node2.

dmag::Float : panel length

FLOWFoil.get_normal — Function

get_normal(d, dmag)

Get unit normal to panel.

Arguments:

d::Vector{Float} : vector from node1 to node2.

dmag::Float : panel length

FLOWFoil.get_distances — Function

get_distances(node1, node2, point)

Get vectors and magnitudes for panel and between nodes and validation points.

Arguments:

node1::Array{Float} : [x y] position of node1.

node2::Array{Float} : [x y] position of node2.

point::Array{Float} : [x y] position of evaluation point.

Returns:

r1::Vector{Float} : vector from node1 to evaluation point.

r1mag::Float : distance from node1 to evaluation point.

r2::Vector{Float} : vector from node2 to evaluation point.

r2mag::Float : distance from node2 to evaluation point.

d::Vector{Float} : vector from node1 to node2.

dmag::Float : panel length

FLOWFoil.get_orientation — Function

get_orientation(node1, node2, point)

Get angles between panel and evaluation point, ln of distances from nodes to evaluation point, and evaluation point position relative to panel.

Arguments:

node1::Array{Float} : [x y] position of node1.

node2::Array{Float} : [x y] position of node2.

point::Array{Float} : [x y] position of evaluation point.

Returns:

theta1::Float : Angle between panel and evaluation point, centered at node1.

theta2::Float : Angle between panel and evaluation point, centered at node2.

ln1::Float : Natural log of distance from node1 to evaluation point.

ln2::Float : Natural log of distance from node2 to evaluation point.

h::Float : Distance from panel to evaluation in panel normal direction.

a::Float : Distance from node1 to evaluation in panel tangent direction.

Singularity Influences

FLOWFoil.get_psibargamma — Function

get_psibargamma(theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\overline{\Psi}^\gamma$

Arguments:

theta1::Float : angle between panel and vector from node1 to evaluation point

theta2::Float : angle between panel and vector from node2 to evaluation point

ln1::Float : value of ln(rmag1), which may be that or 0.0, depening on evaluation point location

ln2::Float : value of ln(rmag2), which may be that or 0.0, depening on evaluation point location

dmag::Float : panel length

h::Float : height of right triangle with hypontenuse, r1, and base, a, colinear with panel.

a::Float : length of base of right triangle with height, h, and hypontenuse, r1.

FLOWFoil.get_psitildegamma — Function

get_psitildegamma(psibargamma, r1mag, r2mag, theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\widetilde{\Psi}^\gamma$

Arguments:

psibargamma::Float : value of $\overline{\Psi}^\gamma$

r1mag::Float : distance from node1 to evaluation point

r2mag::Float : distance from node2 to evaluation point

theta1::Float : angle between panel and vector from node1 to evaluation point

theta2::Float : angle between panel and vector from node2 to evaluation point

ln1::Float : value of ln(rmag1), which may be that or 0.0, depening on evaluation point location

ln2::Float : value of ln(rmag2), which may be that or 0.0, depening on evaluation point location

dmag::Float : panel length

h::Float : height of right triangle with hypontenuse, r1, and base, a, colinear with panel.

a::Float : length of base of right triangle with height, h, and hypontenuse, r1.

FLOWFoil.get_vortex_influence — Function

get_vortex_influence(node1, node2, point)

Calculate vortex influence coefficients on the evaluation point from the panel between node1 and node2.

Arguments:

node1::Array{Float}(2) : [x y] location of node1

node2::Array{Float}(2) : [x y] location of node2

point::Array{Float}(2) : [x y] location of evaluation point

FLOWFoil.get_psibarsigma — Function

get_psibarsigma(theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\overline{\Psi}^\sigma$

Arguments:

theta1::Float : Angle between panel and evaluation point, centered at node1.

theta2::Float : Angle between panel and evaluation point, centered at node2.

ln1::Float : Natural log of distance from node1 to evaluation point.

ln2::Float : Natural log of distance from node2 to evaluation point.

h::Float : Distance from panel to evaluation in panel normal direction.

a::Float : Distance from node1 to evaluation in panel tangent direction.

FLOWFoil.get_psitildesigma — Function

get_psitildesigma(psibargamma, r1mag, r2mag, theta1, theta2, ln1, ln2, dmag, h, a)

Calculate value of $\widetilde{\Psi}^\sigma$

Arguments:

psibargamma::Float : value of $\overline{\Psi}^\sigma$

r1mag::Float : distance from node1 to evaluation point

r2mag::Float : distance from node2 to evaluation point

theta1::Float : angle between panel and vector from node1 to evaluation point

theta2::Float : angle between panel and vector from node2 to evaluation point

dmag::Float : panel length

h::Float : height of right triangle with hypontenuse, r1, and base, a, colinear with panel.

a::Float : length of base of right triangle with height, h, and hypontenuse, r1.

FLOWFoil.get_source_influence — Function

get_source_influence(node1, node2, point)

Calculate source influence coefficients on the evaluation point from the panel between node1 and node2.

Arguments:

node1::Array{Float}(2) : [x y] location of node1

node2::Array{Float}(2) : [x y] location of node2

point::Array{Float}(2) : [x y] location of evaluation point

Matrix Assembly

FLOWFoil.InviscidSystem — Type

InviscidSystem{TF}

Fields:

vcoeffmat::Array{Float,2} : Vortex Coefficient Matrix used in solution.

bccoeffvec::Array{Float,2} : Boundary Coefficient Vector used in solution.

Ns::Array{Float} : Array of numbers of nodes for each airfoil in the system.

FLOWFoil.assemble_vortex_coefficients — Function

assemble_vortex_coefficients(meshi, meshj)

Assemble matrix of vortex strength coefficients.

This function only assembles the NxM portion of the system influence coefficient matrix associated with the M-1 panels of meshj acting on the N nodes of meshi. It does not include the kutta condition or the influence of the constant stream function on the airfoil nodes.

Arguments:

meshi::BodyMesh : mesh being influenced.

meshj::BodyMesh : mesh doing the influencing.

trailing_edge_treatment::Bool : flag for whether to treat trailing edge or not (is meshi==meshj?)

FLOWFoil.assemble_vortex_matrix — Function

assemble_vortex_matrix(meshes)

Assemble vortex coefficient matrix with full N+n x N+n system, including Kutta condition, where n is the number of meshes (airfoils) in the system, and N is the total number of nodes between all the meshes.

Arguments:

meshes::Array{BodyMesh} : Mesh System for which to solve.

FLOWFoil.assemble_boundary_conditions — Function

assemble_boundary_conditions(meshes)

Assemble boundary condition vector.

Arguments:

meshes::Array{BodyMesh} : mesh system for which to solve

Returns

psi_inf::Array{Float,2} : Boundary condition array.

FLOWFoil.get_inviscid_system — Function

get_inviscid_system(meshes)

Calculate, then gather the vortex and boundary condition matricies into an InviscidSystem object.

Arguments:

meshes::Array{BodyMesh} : BodyMesh for airfoil to analyze.

Inviscid System Solving

FLOWFoil.solve_inviscid — Function

solve_inviscid(problem)

Solves the inviscid problem.

Arguments:

problem::Problem : Problem to solve. viscous field must be set to false.

Returns:

solution::InviscidSolution

FLOWFoil.solve_inviscid_system — Function

solve_inviscid_system(inviscidsystem, mesh; debug=false)

Solve the InviscidSystem for the vortex and streamfunction strengths.

Outputs the InviscidSolution object which contains the inviscidsystem in the debug object if debug is set to true.

Arguments:

inviscidsystem::InviscidSystem : InviscidSystem to solve.

mesh::BodyMesh : BodyMesh defining geometry (to put into solution object)

Keyword Arguments:

debug::Bool = false : flag to indicate whether or not to output all the system details.

Returns:

solution::InviscidSolution

Post Processing

FLOWFoil.get_vortex_magnitudes — Function

get_vortex_magnitudes(inviscid_solution,angleofattack)

Calculate the vortex strength magnitudes at the airfoil nodes for a given angle of attack.

Arguments:

inviscid_solution::InviscidSolution : the inviscid solution from which to find the vortex magnitudes.

angleofattack::Float : the angle of attack in degrees.

FLOWFoil.get_stream_grid_value — Function

get_stream_grid_value(gammas, nodes, point, vinf, alpha, blunt_te=false, txp=0.0, tdp=0.0)

For a given point, calculate the total stream function value induced by an airfoil.

Arguments:

gammas::Array{Float} : Array of vortex strength magnitudes at each of the nodes.

sigmas::Array{Float} : Array of source strength magnitudes at each of the nodes.

nodes::Array{Array{Float}} : Array of node locations

point::Array{Float} : [x z] location of the point in question

vinf::Float : Freestream magnitude

alpah::Float : Angle of attack in degrees

blunt_te::Bool : Flag whether airfoil has a blunt trailing edge or not

txp::Float : txp field from mesh object where nodes came from

tdp::Float : tdp field from mesh object where nodes came from

FLOWFoil.get_gamma_magnitudes — Function

get_gamma_magnitudes(panelgammas, angleofattack)

Calculate vortex strength magnitudes for each of the nodes.

Arguements:

panelgammas::Array{Float,2} : panelgammas field from solution object.

angleofattack::Float : Angle of attack in degrees.

Convenience Functions

FLOWFoil.cosinespacing — Function

function cosinespacing(N=180)

Calculate N cosine spaced points.

Arguments:

N::Int : Number of points

FLOWFoil.split_upper_lower — Function

split_upper_lower(x, z)

Split the upper and lower halves of the airfoil coordinates. Assumes odd number of coordinates (leading edge repeated).

Arguments:

x::Array{Float} : Array of x coordinates

z::Array{Float} : Array of z coordinates

Returns:

xu::Array{Float} : Array of upper half of x coordinates

xl::Array{Float} : Array of lower half of x coordinates

zu::Array{Float} : Array of upper half of z coordinates

zl::Array{Float} : Array of lower half of z coordinates

FLOWFoil.normalize_airfoil! — Function

normalize_airfoil!(x, z)

Normalize airfoil to unit chord and shift leading edge to zero. Adjusts coordinates in place.

Arguments:

x::Array{Float} : Array of x coordinates

z::Array{Float} : Array of z coordinates

FLOWFoil.thickness — Function

thickness(x, maxthick; blunt_TE=false)

Compute thickness at a given chord-normalized x-position by NACA 4-series thickness equations.

Arguments:

x::Float : x position along chordline

maxthick::Float : Maximum thickness value

Keyword Arguments:

blunt_TE::Bool : Flag whether trailing edge is blunt or not

FLOWFoil.camber — Function

camber(x, maxcamber, camberpose)

Compute camber at a given chord-normalized x-position by NACA 4-series camber equations.

Arguments:

x::Float : x position along chordline

maxcamber::Float64 : Maximum camber value

camberpose::Float64 : Position of maximum camber